+++ .github/workflows/library.yml
@@ -44,7 +44,7 @@ jobs:
           git stash
           git checkout $GITHUB_HEAD_REF
           git pull
-          pre-commit run --all-files
+          pre-commit run --all-files || true
           git commit -m "Fix pre-commit failures" -a
 
           echo "##[set-output name=complete;]$(echo 1)"
