+++ setup.py
@@ -173,6 +173,7 @@ package_data = {
         "lib/ref/*",
         "lib/ref/experimental/*",
         "lib/tutes/*",
+        "lib/tutes/gamma/*",
         "trees/*",
     ],
     "pymedphys.docs._build.html._static": [
@@ -338,7 +339,7 @@ setup_kwargs = {
     "name": "pymedphys",
     "version": "0.38.0",
     "description": "Medical Physics library",
-    "long_description": "|logo|\n\n.. |logo| image:: https://github.com/pymedphys/pymedphys/raw/ca501275227f190a77e641a75af925d9070952b6/lib/pymedphys/docs/_static/pymedphys_title.svg\n    :target: https://docs.pymedphys.com/\n\n.. START_OF_DOCS_IMPORT\n\n**A community effort to develop an open standard library for Medical Physics\nin Python. We build high quality, transparent software together via peer review\nand open source distribution. Open code is better science.**\n\n|build| |pypi| |python| |license|\n\n.. |build| image:: https://img.shields.io/github/workflow/status/pymedphys/pymedphys/Library\n    :target: https://github.com/pymedphys/pymedphys/actions\n\n.. |pypi| image:: https://img.shields.io/pypi/v/pymedphys\n    :target: https://pypi.org/project/pymedphys/\n\n.. |python| image:: https://img.shields.io/pypi/pyversions/pymedphys\n    :target: https://pypi.org/project/pymedphys/\n\n.. |license| image:: https://img.shields.io/pypi/l/pymedphys\n    :target: https://choosealicense.com/licenses/apache-2.0/\n\n\nWhat is PyMedPhys?\n==================\n\nPyMedPhys is an open-source Medical Physics python library built by an open\ncommunity that values and prioritises code sharing, review, improvement, and\nlearning from each other. It is inspired by the collaborative work of our\nphysics peers in astronomy and the `Astropy Project`_. PyMedPhys is available\non `PyPI`_ and `GitHub`_.\n\n.. _`Astropy Project`: http://www.astropy.org/\n.. _`PyPI`: https://pypi.org/project/pymedphys/\n.. _`GitHub`: https://github.com/pymedphys/pymedphys\n\nBeta level of development\n*************************\n\nPyMedPhys is currently within the ``beta`` stage of its life-cycle. It will\nstay in this stage until the version number leaves ``0.x.x`` and enters\n``1.x.x``. While PyMedPhys is in ``beta`` stage, **no API is guaranteed to be\nstable from one release to the next.** In fact, it is very likely that the\nentire API will change multiple times before a ``1.0.0`` release. In practice,\nthis means that upgrading ``pymedphys`` to a new version will possibly break\nany code that was using the old version of pymedphys. We try to be abreast of\nthis by providing details of any breaking changes from one release to the next\nwithin the `Release Notes`_.\n\nCommunity\n**************\n\nPyMedPhys has a `Discourse community <https://pymedphys.discourse.group/>`_\nto both help you find your feet using PyMedPhys and to facilitate collaboration\nand general discussion. Please reach out over there and we'd love to get to\nknow you!\n\nDocumentation\n=============\n\nPyMedPhys can be installed with:\n\n.. code:: bash\n\n    pip install pymedphys[user]\n\nIf you'd like more details around installing PyMedPhys go to the\n`getting started`_ page within t
[…]
