+++ pyproject.toml
@@ -1,7 +1,7 @@
 [tool.poetry]
 name = "pymedphys"
 version = "0.38.0"
-readme = "lib/pymedphys/docs/README.rst"
+readme = "README.rst"
 description = "Medical Physics library"
 authors = [
     "PyMedPhys Contributors <developers@pymedphys.com>",
