+++ _sources/release-notes.md
@@ -1,3 +1,9 @@
+<!-- DO NOT EDIT THIS FILE! -->
+<!-- This file has been autogenerated by `poetry run pymedphys dev propagate` -->
+<!-- Please instead edit the file found at: -->
+<!--     pymedphys/CHANGELOG.md -->
+<!-- and then run `poetry run pymedphys dev propagate --copies` -->
+
 <!-- markdownlint-disable MD024 MD039 -->
 
 # Release Notes
@@ -11,42 +17,42 @@ This project adheres to
 
 ### New features and enhancements
 
-* The Gamma tool now utilises the `logging` module. The `quiet` parameter for
+- The Gamma tool now utilises the `logging` module. The `quiet` parameter for
   `pymedphys.gamma` is deprecated but its use is non-breaking (for now).
 
 ## [0.38.0]
 
 ### News around this release
 
-* [Phillip Chlap](https://github.com/pchlap) has joined as a pymedphys Maintainer
-* Thanks to [Cedric](https://github.com/cedricbelanger) for reporting and fixing DICOM coordinates bug: [Discourse](https://pymedphys.discourse.group/t/gamma-index-calculation-possible-issues/191/2)
-* Privacy related bug within the experimental components of pymedphys (not stable) has been fixed. See bug fix notes.
+- [Phillip Chlap](https://github.com/pchlap) has joined as a pymedphys Maintainer
+- Thanks to [Cedric](https://github.com/cedricbelanger) for reporting and fixing DICOM coordinates bug: [Discourse](https://pymedphys.discourse.group/t/gamma-index-calculation-possible-issues/191/2)
+- Privacy related bug within the experimental components of pymedphys (not stable) has been fixed. See bug fix notes.
 
 ### Bug Fixes
 
-* Pseudonymise was susceptible to brute-force reversal, mitigated in [Pull Request](https://github.com/pymedphys/pymedphys/pull/1580)
-* DICOM Coordinates
-  * Fix pixel spacing bug [Pull Request](https://github.com/pymedphys/pymedphys/pull/1571)
-* Pinnacle Export Tool
-  * Write DICOM files to be more compatible with other software [Pull Request](https://github.com/pymedphys/pymedphys/pull/1578)
-* Daily Winston-lutz
-  * Small workaround [Pull Request](https://github.com/pymedphys/pymedphys/pull/1565)
+- Pseudonymise was susceptible to brute-force reversal, mitigated in [Pull Request](https://github.com/pymedphys/pymedphys/pull/1580)
+- DICOM Coordinates
+  - Fix pixel spacing bug [Pull Request](https://github.com/pymedphys/pymedphys/pull/1571)
+- Pinnacle Export Tool
+  - Write DICOM files to be more compatible with other software [Pull Request](https://github.com/pymedphys/pymedphys/pull/1578)
+- Daily Winston-lutz
+  - Small workaround [Pull Request](https://github.com/pymedphys/pymedphys/pull/1565)
 
 ### New features and enhancements
 
-* Pinnacle Export Tool
-  * A new CLI option is provided to append the patient MRN to the output path for convenience. Thanks [crewso](https://github.com/crcrewso) for submitting this [Pull Request](https://github.com/pymedphys/pymedphys/pull/1608)
+- Pinnacle Export Tool
+  - A new CLI option is provided to append the patient MRN to the output path for convenience. Thanks [crewso](https://github.com/crcrewso) for submitting this [Pull Request](https://github.com/pymedphys/pymedphys/pull/1608)
 
 ## [0.37.1]
 
 ### Bug fixes
 
-* Make it so that Mephysto files that have non-unicode characters can still
+- Make it so that Mephysto files that have non-unicode characters can still
   be opened.
[…]
