+++ _sources/README.rst
@@ -1,3 +1,14 @@
+..
+    DO NOT EDIT THIS FILE!
+..
+    This file has been autogenerated by `poetry run pymedphys dev propagate`
+..
+    Please instead edit the file found at:
+..
+        pymedphys/README.rst
+..
+    and then run `poetry run pymedphys dev propagate --copies`
+
 |logo|
 
 .. |logo| image:: https://github.com/pymedphys/pymedphys/raw/ca501275227f190a77e641a75af925d9070952b6/lib/pymedphys/docs/_static/pymedphys_title.svg
@@ -85,8 +96,8 @@ The PyMedPhys documentation is split into four categories:
 Our Team
 ========
 
-PyMedPhys is what it is today due to its maintainers and contributors, both past
-and present. Here is our team.
+PyMedPhys is what it is today due to its maintainers and contributors, both
+past and present. Here is our team.
 
 Maintainers
 ***********
