+++ lib/pymedphys/docs/contrib/get-started.rst
@@ -1,0 +1,3 @@
+=========================
+Get Started Contributing
+=========================
