+++ lib/pymedphys/docs/users/howto/index.md
@@ -4,7 +4,7 @@
 ```{toctree}
 :maxdepth: 2
 
-get-started
+../get-started
 gamma/index
 add-a-linac
 tunnels/index
