+++ lib/pymedphys/docs/users/howto/index.md
@@ -4,7 +4,6 @@
 ```{toctree}
 :maxdepth: 2
 
-../get-started
 gamma/index
 add-a-linac
 tunnels/index
