+++ .readthedocs.yml
@@ -1,0 +1,22 @@
+version: 2
+
+build:
+  os: ubuntu-22.04
+  tools:
+    python: "3.8"
+  jobs:
+    pre_build:
+      # Generate the Sphinx configuration for this Jupyter Book so it builds.
+      - "jupyter-book config sphinx lib/pymedphys/docs/"
+
+python:
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+      - sphinx
+
+sphinx:
+  builder: html
+  fail_on_warning: false
+  configuration: lib/pymedphys/docs/conf.py
