### 2022-06-15 c00d5bd
Re-include contrib setups
