### 2022-06-15 #4470 @danasolav
Hi @arfon,
I could not understand from the readme file what this software does. Medical physics is extremely broad. According to the paper it looks like it focuses on radiation, but the paper also lacks some key information that it should contain. Perhaps the authors can revise the readme and paper to be more specific and informative?
In any case, radiation is not my area of expertise, so it looks like I'm going to have to decline this one.
