+++ .readthedocs.yml
@@ -14,7 +14,7 @@ python:
     - method: pip
       path: .
       extra_requirements:
-      - sphinx
+      - docs
 
 sphinx:
   builder: html
