### 2022-06-15 df3e0c3
reword changelog description. add release notes to index re-add tips & tricks
