+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # Release Notes
 
-All notable changes to this project will be documented in this file.
+All notable changes to are documented here.
 
 This project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
