+++ .github/workflows/conflicts.yml
@@ -3,6 +3,8 @@ on:
 
 jobs:
   FixPropagateConflicts:
+    if: false
+
     runs-on: ubuntu-latest
     timeout-minutes: 60
 
