+++ .github/workflows/paper.yml
@@ -8,6 +8,7 @@ on:
 
 jobs:
   paper:
+    if: false
     runs-on: ubuntu-latest
     name: Paper Draft
     steps:
