+++ .github/workflows/mosaiq-db-tests.yml
@@ -10,6 +10,7 @@ jobs:
   # =============================================================================
 
   TestMosaiqDb:
+    if: false
     runs-on: ubuntu-latest
     timeout-minutes: 60
 
