+++ pyproject.toml
@@ -48,7 +48,7 @@ PyYAML = { version = "*", optional = true }          # groups = ["user", "dev"]
 requests = { version = "*", optional = true }        # groups = ["user", "dev"]
 python-dateutil = { version = "*", optional = true } # groups = ["user", "dev", "tests"]
 
-matplotlib = { version = "*", optional = true } # groups = ["user", "dev"]
+matplotlib = { version = "*", optional = true } # groups = ["user", "dev", "docs"]
 scipy = { version = "*", optional = true }      # groups = ["user", "dev"]
 
 xarray = { version = "*", optional = true } # groups = ["user", "dev"]
@@ -219,7 +219,7 @@ dev = [
     "xmltodict",
 ]
 dicom = ["pydicom", "pynetdicom"]
-docs = ["jupyter-book", "networkx", "sphinx-argparse", "sphinx-book-theme"]
+docs = ["jupyter-book", "matplotlib", "networkx", "sphinx-argparse", "sphinx-book-theme"]
 doctests = ["black", "pylinac", "sphinx-book-theme", "tabulate"]
 icom = ["numpy"]
 mosaiq = ["pandas", "pymssql", "scikit-learn", "sqlalchemy", "toml"]
