### 2022-06-15 41fce96
remove excess matrix
