+++ .github/workflows/library.yml
@@ -13,6 +13,8 @@ jobs:
   # =============================================================================
 
   PreCommit:
+    if: false
+
     runs-on: ubuntu-latest
     timeout-minutes: 60
 
@@ -120,10 +122,10 @@ jobs:
           # See https://jupyterbook.org/advanced/advanced.html#working-on-windows
           - os: "ubuntu"
             python-version: "3.10"
-          - os: "windows"
-            python-version: "3.10"
-          - os: "macos"
-            python-version: 3.8
+          # - os: "windows"
+          #   python-version: "3.10"
+          # - os: "macos"
+          #   python-version: 3.8
 
     runs-on: ${{ matrix.os }}-latest
     timeout-minutes: 60
