### 2022-06-15 830bf11
add cli, organise tech ref
