+++ lib/pymedphys/docs/users/ref/index.md
@@ -1,0 +1,9 @@
+
+# Technical Reference
+
+```{toctree}
+:maxdepth: 2
+
+lib-ref/index
+cli-ref/index
+```
