+++ lib/pymedphys/docs/users/ref/lib-ref/index.rst
@@ -1,6 +1,6 @@
-=========
-Reference
-=========
+===================
+Library Reference
+===================
 
 PyMedPhys has a range of modules that are able to be exposed from within Python
 itself. Below is the documentation for a range of these libraries.
