+++ pyproject.toml
@@ -73,11 +73,11 @@ imageio = { version = "*", optional = true }           # groups = ["user", "dev"
 pylibjpeg-libjpeg = { version = "*", optional = true } # groups = ["user", "dev"]
 
 # NumPy's lower bound here is for typing support
-numpy = { version = ">=1.20.2", optional = true } # groups = ["user", "dev", "icom"]
+numpy = { version = ">=1.20.2", optional = true } # groups = ["user", "dev", "icom", "docs"]
 
 # The following lower bounded packages are due to certain APIs utilised
 pandas = { version = ">=1.0.0", optional = true }  # groups = ["user", "dev", "mosaiq"]
-pydicom = { version = ">=2.0.0", optional = true } # groups = ["user", "dev", "dicom"]
+pydicom = { version = ">=2.0.0", optional = true } # groups = ["user", "dev", "dicom", "docs"]
 
 
 # The following lower bounds are so that the versions with wheels are utilised
