+++ requirements-docs.txt
@@ -1,0 +1,135 @@
+alabaster==0.7.12; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
+anyio==3.6.1; python_full_version >= "3.6.2" and python_version >= "3.7"
+appnope==0.1.3; sys_platform == "darwin" and python_version >= "3.8" and platform_system == "Darwin"
+argon2-cffi-bindings==21.2.0; python_version >= "3.7"
+argon2-cffi==21.3.0; python_version >= "3.7"
+asttokens==2.0.5; python_version >= "3.8"
+attrs==21.4.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
+babel==2.10.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
+backcall==0.2.0; python_version >= "3.8"
+beautifulsoup4==4.11.1; python_full_version >= "3.6.0" and python_version >= "3.7"
+bleach==5.0.0; python_version >= "3.7"
+certifi==2022.5.18.1; python_version >= "3.7" and python_version < "4" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7")
+cffi==1.15.0; python_full_version >= "3.7.0" and python_version >= "3.7" and implementation_name == "pypy"
+charset-normalizer==2.0.12; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.5.0" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7")
+click==8.1.3; python_version >= "3.7" and python_version < "4.0"
+colorama==0.4.4; python_version >= "3.8" and python_full_version < "3.0.0" and platform_system == "Windows" and sys_platform == "win32" or python_full_version >= "3.5.0" and platform_system == "Windows" and sys_platform == "win32" and python_version >= "3.8"
+cycler==0.11.0; python_version >= "3.7"
+debugpy==1.6.0; python_version >= "3.7"
+decorator==5.1.1; python_version >= "3.8"
+defusedxml==0.7.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
+docutils==0.17.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
+entrypoints==0.4; python_full_version >= "3.7.0" and python_version >= "3.7"
+executing==0.8.3; python_version >= "3.8"
+fastjsonschema==2.15.3; python_version >= "3.7"
+fonttools==4.33.3; python_version >= "3.7"
+gitdb==4.0.9; python_version >= "3.7"
+gitpython==3.1.27; python_version >= "3.7"
[…]
