+++ .readthedocs.yml
@@ -11,10 +11,9 @@ build:
 
 python:
   install:
+    - requirements: requirements-docs.txt
     - method: pip
       path: .
-      extra_requirements:
-      - docs
 
 sphinx:
   builder: html
