+++ pyproject.toml
@@ -38,7 +38,7 @@ typing-extensions = "*"
 # ------- #
 # Library #
 # ------- #
-tqdm = { version = "*", optional = true }     # groups = ["user", "dev", "tests"]
+tqdm = { version = "*", optional = true }     # groups = ["user", "dev", "tests", "docs"]
 attrs = { version = "*", optional = true }    # groups = ["user", "dev"]
 watchdog = { version = "*", optional = true } # groups = ["user", "dev"]
 
@@ -219,7 +219,16 @@ dev = [
     "xmltodict",
 ]
 dicom = ["pydicom", "pynetdicom"]
-docs = ["jupyter-book", "matplotlib", "networkx", "sphinx-argparse", "sphinx-book-theme"]
+docs = [
+    "jupyter-book",
+    "matplotlib",
+    "networkx",
+    "numpy",
+    "pydicom",
+    "sphinx-argparse",
+    "sphinx-book-theme",
+    "tqdm",
+]
 doctests = ["black", "pylinac", "sphinx-book-theme", "tabulate"]
 icom = ["numpy"]
 mosaiq = ["pandas", "pymssql", "scikit-learn", "sqlalchemy", "toml"]
