### 2022-06-15 6794002
Flesh out tutorials a little Rename gamma tute with more specificity
