+++ lib/pymedphys/docs/users/tutes/index.rst
@@ -2,7 +2,10 @@
 Tutorials
 =========
 
-There are not tutorials yet! =(
+.. toctree::
+    :maxdepth: 2
+
+    gamma/index
 
 The goal for these tutorials is to have a suite of Jupyter Notebooks that can
 be interactively run within the documentation to help Medical Physicists who
