+++ lib/pymedphys/docs/users/tutes/gamma/index.rst
@@ -5,4 +5,4 @@ Gamma
 .. toctree::
     :maxdepth: 2
 
-    gamma_tutorial
+    gamma_shell_accelerations
