### 2022-06-15 0688df8
reorganise get-started docs fix "cmd lexer not found" in quick-start
