+++ lib/pymedphys/docs/users/get-started/cli.rst
@@ -1,6 +1,6 @@
-========
-Overview
-========
+============================
+Using the CLI
+============================
 
 PyMedPhys exposes a range of its library functions to the command line as a
 command line interface (CLI). This provides the benefit of being able to call
@@ -11,4 +11,4 @@ PyMedPhys functions from a range of locations such as:
 * ``.bat`` and ``.sh`` files
 
 You can find the technical reference for the command line tools
-:doc:`here<ref/index>`.
+:doc:`here<../users/ref/cli-ref/index>`.
