### 2022-06-15 a3c615b
revert tests removal, fix docs dep
