+++ setup.py
@@ -81,7 +81,6 @@ packages = [
     "pymedphys.cli",
     "pymedphys.cli.experimental",
     "pymedphys.docs",
-    "pymedphys.docs._build.html._static",
     "pymedphys.experimental",
     "pymedphys.tests",
     "pymedphys.tests.coordinates",
@@ -120,42 +119,6 @@ package_data = {
     "pymedphys._experimental.serviceplans": ["templates/*"],
     "pymedphys._experimental.streamlit.apps": ["data/*"],
     "pymedphys.docs": [
-        "_build/html/*",
-        "_build/html/_images/*",
-        "_build/html/_modules/*",
-        "_build/html/_modules/pymedphys/_dicom/*",
-        "_build/html/_modules/pymedphys/_experimental/pinnacle/*",
-        "_build/html/_modules/pymedphys/_mosaiq/*",
-        "_build/html/_panels_static/*",
-        "_build/html/_sources/*",
-        "_build/html/_sources/cli/*",
-        "_build/html/_sources/cli/ref/*",
-        "_build/html/_sources/contrib/*",
-        "_build/html/_sources/contrib/other/*",
-        "_build/html/_sources/contrib/setups/*",
-        "_build/html/_sources/general/*",
-        "_build/html/_sources/lib/*",
-        "_build/html/_sources/lib/background/*",
-        "_build/html/_sources/lib/howto/*",
-        "_build/html/_sources/lib/howto/gamma/*",
-        "_build/html/_sources/lib/howto/tunnels/*",
-        "_build/html/_sources/lib/ref/*",
-        "_build/html/_sources/lib/ref/experimental/*",
-        "_build/html/_sources/lib/tutes/*",
-        "_build/html/cli/*",
-        "_build/html/cli/ref/*",
-        "_build/html/contrib/*",
-        "_build/html/contrib/other/*",
-        "_build/html/contrib/setups/*",
-        "_build/html/general/*",
-        "_build/html/lib/*",
-        "_build/html/lib/background/*",
-        "_build/html/lib/howto/*",
-        "_build/html/lib/howto/gamma/*",
-        "_build/html/lib/howto/tunnels/*",
-        "_build/html/lib/ref/*",
-        "_build/html/lib/ref/experimental/*",
-        "_build/html/lib/tutes/*",
         "_static/*",
         "cli/*",
         "cli/ref/*",
@@ -175,14 +138,6 @@ package_data = {
         "lib/tutes/*",
         "lib/tutes/gamma/*",
         "trees/*",
-    ],
-    "pymedphys.docs._build.html._static": [
-        "css/*",
-        "images/*",
-        "js/*",
-        "vendor/fontawesome/5.13.0/*",
-        "vendor/fontawesome/5.13.0/css/*",
-        "vendor/fontawesome/5.13.0/webfonts/*",
     ],
     "pymedphys.tests.dicom": ["data/rtplan/*", "scratch/*"],
     "pymedphys.tests.e2e": [
@@ -241,6 +196,7 @@ extras_require = {
         "streamlit-ace==0.1.0",
         "pylinac==2.3.2",
         "reportlab>=3.6",
+        "ipython>=7.31.1",
         "Pillow>=9.0.0",
         "sphinx-argparse",
         "sphinx-book-theme",
@@ -267,8 +223,11 @@ extras_require = {
     "docs": [
         "tqdm",
         "matplotlib",
+        "scipy",
+        "toml",
         "numpy>=1.20.2",
         "pydicom>=2.0.0",
+        "ipython>=7.31.1",
         "sphinx-argparse",
         "sphinx-book-theme",
         "jupyter-book>=0.12.0",
