+++ .github/workflows/library.yml
@@ -13,7 +13,7 @@ jobs:
   # =============================================================================
 
   PreCommit:
-    if: false
+    # if: false
 
     runs-on: ubuntu-latest
     timeout-minutes: 60
@@ -66,21 +66,17 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        os: [
-            "ubuntu",
-            # "macos",
-            # "windows",
-          ]
+        os: ["ubuntu", "macos", "windows"]
         task: [
-            # "tests",
+            "tests",
             "docs",
-            # "stackoverflow",
-            # "updates",
-            # "cypress",
-            # "slow",
+            "stackoverflow",
+            "updates",
+            "cypress",
+            "slow",
             # "binary",  # Skip for now
-            # "pyright",
-            # "propagate",
+            "pyright",
+            "propagate",
           ]
 
         exclude:
@@ -122,10 +118,10 @@ jobs:
           # See https://jupyterbook.org/advanced/advanced.html#working-on-windows
           - os: "ubuntu"
             python-version: "3.10"
-          # - os: "windows"
-          #   python-version: "3.10"
-          # - os: "macos"
-          #   python-version: 3.8
+          - os: "windows"
+            python-version: "3.10"
+          - os: "macos"
+            python-version: 3.8
 
     runs-on: ${{ matrix.os }}-latest
     timeout-minutes: 60
