+++ .github/workflows/mosaiq-db-tests.yml
@@ -10,7 +10,8 @@ jobs:
   # =============================================================================
 
   TestMosaiqDb:
-    if: false
+    # if: false
+
     runs-on: ubuntu-latest
     timeout-minutes: 60
 
