+++ .github/workflows/conflicts.yml
@@ -3,7 +3,7 @@ on:
 
 jobs:
   FixPropagateConflicts:
-    if: false
+    # if: false
 
     runs-on: ubuntu-latest
     timeout-minutes: 60
