+++ lib/pymedphys/docs/users/howto/gamma/index.rst
@@ -5,5 +5,6 @@ Gamma
 .. toctree::
     :maxdepth: 2
 
+    1D-from-csv
     effect-of-noise
     from-dicom
