+++ lib/pymedphys/docs/general/index.rst
@@ -1,11 +1,0 @@
-======================
-Overview
-======================
-
-This section contains the PyMedPhys :doc:`Release Notes<../release-notes>`,
-along with with additional documentation that may be of interest to any
-PyMedPhys visitor.
-
-- :doc:`Release Notes<../release-notes>`
-- :doc:`File Structure<file-structure>`
-- :doc:`Benefits of Open Source<open-source-benefits>`
