+++ lib/pymedphys/docs/users/ref/lib-ref/index.rst
@@ -8,7 +8,7 @@ itself. Below is the documentation for a range of these libraries.
 .. toctree::
     :maxdepth: 2
 
-    ../../genindex
+    ../../../genindex
     dicom
     gamma
     mosaiq
