+++ lib/pymedphys/docs/contrib/info/index.md
@@ -1,0 +1,9 @@
+
+# Repository Information
+
+```{toctree}
+:maxdepth: 2
+
+file-structure
+documentation
+```
