+++ lib/pymedphys/docs/contrib/get-started.rst
@@ -1,3 +1,0 @@
-=========================
-Get Started Contributing
-=========================
