+++ setup.py
@@ -120,29 +120,29 @@ package_data = {
     "pymedphys._experimental.streamlit.apps": ["data/*"],
     "pymedphys.docs": [
         "_static/*",
-        "cli/*",
-        "cli/ref/*",
         "contrib/*",
-        "contrib/other/*",
+        "contrib/info/*",
         "contrib/setups/*",
-        "general/*",
+        "contrib/tips-tricks/*",
         "img/*",
-        "lib/*",
-        "lib/background/*",
-        "lib/howto/*",
-        "lib/howto/gamma/*",
-        "lib/howto/tunnels/*",
-        "lib/howto/tunnels/img/*",
-        "lib/ref/*",
-        "lib/ref/experimental/*",
-        "lib/tutes/*",
-        "lib/tutes/gamma/*",
         "trees/*",
+        "users/background/*",
+        "users/get-started/*",
+        "users/howto/*",
+        "users/howto/gamma/*",
+        "users/howto/tunnels/*",
+        "users/howto/tunnels/img/*",
+        "users/ref/*",
+        "users/ref/cli-ref/*",
+        "users/ref/lib-ref/*",
+        "users/ref/lib-ref/experimental/*",
+        "users/tutes/*",
+        "users/tutes/gamma/*",
     ],
     "pymedphys.tests.dicom": ["data/rtplan/*", "scratch/*"],
     "pymedphys.tests.e2e": [
         "cypress/*",
-        "cypress/fixtures/.gitignore",
+        "cypress/fixtures/*",
         "cypress/integration/streamlit/*",
         "cypress/plugins/*",
         "cypress/support/*",
@@ -302,7 +302,7 @@ setup_kwargs = {
     "name": "pymedphys",
     "version": "0.38.0",
     "description": "Medical Physics library",
-    "long_description": "|logo|\n\n.. |logo| image:: https://github.com/pymedphys/pymedphys/raw/ca501275227f190a77e641a75af925d9070952b6/lib/pymedphys/docs/_static/pymedphys_title.svg\n    :target: https://docs.pymedphys.com/\n\n.. START_OF_DOCS_IMPORT\n\n**A community effort to develop an open standard library for Medical Physics\nin Python. We build high quality, transparent software together via peer review\nand open source distribution. Open code is better science.**\n\n|build| |pypi| |python| |license|\n\n.. |build| image:: https://img.shields.io/github/workflow/status/pymedphys/pymedphys/Library\n    :target: https://github.com/pymedphys/pymedphys/actions\n\n.. |pypi| image:: https://img.shields.io/pypi/v/pymedphys\n    :target: https://pypi.org/project/pymedphys/\n\n.. |python| image:: https://img.shields.io/pypi/pyversions/pymedphys\n    :target: https://pypi.org/project/pymedphys/\n\n.. |license| image:: https://img.shields.io/pypi/l/pymedphys\n    :target: https://choosealicense.com/licenses/apache-2.0/\n\n\nWhat is PyMedPhys?\n==================\n\nPyMedPhys is an open-source Medical Physics python library built by an open\ncommunity that values and prioritises code sharing, review, improvement, and\nlearning from each other. It is inspired by the collaborative work of our\nphysics peers in astronomy and the `Astropy Project`_. PyMedPhys is available\non `PyPI`_ and `GitHub`_.\n\n.. _`Astropy Project`: http://www.astropy.org/\n.. _`PyPI`: https://pypi.org/project/pymedphys/\n.. _`GitHub`: https://github.com/pymedphys/pymedphys\n\nBeta level of development\n*************************\n\nPyMedPhys is curre
[…]
