+++ setup.py
@@ -133,9 +133,9 @@ package_data = {
         "users/howto/tunnels/*",
         "users/howto/tunnels/img/*",
         "users/ref/*",
-        "users/ref/cli-ref/*",
-        "users/ref/lib-ref/*",
-        "users/ref/lib-ref/experimental/*",
+        "users/ref/-ref/*",
+        "users/ref/lib/*",
+        "users/ref/lib/experimental/*",
         "users/tutes/*",
         "users/tutes/gamma/*",
     ],
