+++ lib/pymedphys/docs/users/tutes/index.rst
@@ -1,23 +1,0 @@
-=========
-Tutorials
-=========
-
-.. toctree::
-    :maxdepth: 2
-
-    gamma/index
-
-The goal for these tutorials is to have a suite of Jupyter Notebooks that can
-be interactively run within the documentation to help Medical Physicists who
-have no Python experience begin to have success using the tools within the
-Medical Physics ecosystem. To build a suite of tutorials like this is a big
-task, and most of the developers of PyMedPhys are focused on developing
-PyMedPhys itself. Are you reading this and thinking that you might be able to
-manage and facilitate the creation of a suite of Notebooks like this? You
-certainly don't have to create them all by yourself! Instead, just take the
-task by the horns, design the overall tutorial flow/course, and find people who
-can build each component.
-
-If you would like to consider taking this on, please reach out by sending me
-(@SimonBiggs) a private message on the Discourse group
-(https://pymedphys.discourse.group/u/simonbiggs/summary).
