+++ lib/pymedphys/docs/users/howto/gamma/index.rst
@@ -5,6 +5,7 @@ Gamma
 .. toctree::
     :maxdepth: 2
 
+    speed-up
     1D-from-csv
     effect-of-noise
     from-dicom
