+++ lib/pymedphys/docs/contrib/info/index.md
@@ -5,5 +5,5 @@
 :maxdepth: 2
 
 file-structure
-documentation
+docs-guide
 ```
