+++ lib/pymedphys/docs/users/tutes/gamma/index.rst
@@ -1,8 +1,0 @@
-=====
-Gamma
-=====
-
-.. toctree::
-    :maxdepth: 2
-
-    gamma_shell_accelerations
