+++ lib/pymedphys/docs/index.rst
@@ -23,6 +23,7 @@
     overview
     release-notes
 
+
 .. toctree::
     :hidden:
     :maxdepth: 3
