+++ lib/pymedphys/docs/index.rst
@@ -16,6 +16,7 @@
 .. _`Contributors Guide`: contrib/index.html
 .. _`General`: general/index.html
 
+
 .. toctree::
     :hidden:
     :maxdepth: 3
