### 2022-06-16 pymedphys/pymedphys#1667 @sjswerdloff
The feedback was that they could not understand from the readme file what this software does.
I looked at a JOSS paper at (pseudo) random, and it was for software that calculated a very particular approach to solving a Riemann Problem.
If the readme needs to describe "when I use this software I solve this specific problem with it", then we might want to list what specific tasks people use it for.

Just glancing over the list of directories in the library subdirectory, and trying to describe it to someone who is technical, but not medical, it seems like

PyMedPhys is used to convert data from a variety of proprietary formats in to either an internal/common representation, or in to a standard format, and then provides the means for performing a variety of common Quality Assurance tasks.   The QA task most commonly performed with the software is the calculation of the Gamma index, which
 _quantifies the difference between measured and calculated dose distributions on a point-by-point basis in terms of both dose and distance to agreement (DTA) differences_

I'll get over my profound sense of personal loss </sarcasm>at not having pseudonymisation listed in the Summary.
In other words, I propose we put something in there that can be reviewed by someone who can use a search engine and can read Python, and not try to worry too much about how much we don't tell people about what it can do.

The last time I tried to describe the breadth of what I do (and what I can do) to a management type, I got the response "you are a freak of nature".   There is a warm place in my heart for that person, but it also makes the point that if you need to explain to someone what you do in twenty five words or less at a party, and they've already had a beverage, something simple like "Software for Radiation Therapy for Cancer Care" is good enough, and if they want to get more specific, they need to get another beverage and pull up a chair.  I think we need to try to get to a level of specificity without going so deep that they need to actually *understand* Medical Physics.
