### 2022-06-16 2b01670
simplify propagate, offload docs copies to be run prior to docs build
