+++ lib/pymedphys/_dev/propagate.py
@@ -16,7 +16,6 @@ import json
 import pathlib
 import re
 import subprocess
-import tarfile
 import textwrap
 from typing import List, Optional
 
@@ -29,12 +28,7 @@ POETRY_LOCK_PATH = REPO_ROOT.joinpath("poetry.lock")
 PYPROJECT_TOML_HASH_PATH = REPO_ROOT.joinpath("pyproject.hash")
 
 LIBRARY_PATH = REPO_ROOT.joinpath("lib", "pymedphys")
-DOCS_PATH = LIBRARY_PATH.joinpath("docs")
-
 VERSION_PATH = LIBRARY_PATH.joinpath("_version.py")
-
-DIST_DIR = REPO_ROOT.joinpath("dist")
-SETUP_PY = REPO_ROOT.joinpath("setup.py")
 
 REQUIREMENTS_CONFIG = (
     # Extras | Filename | Include PyMedPhys | Make it an editable dev install
@@ -47,22 +41,6 @@ REQUIREMENTS_CONFIG = (
     (["docs"], "requirements-docs.txt", False, None),
 )
 
-ROOT_README = REPO_ROOT.joinpath("README.rst")
-DOCS_README = DOCS_PATH.joinpath("README.rst")
-
-DOCS_CHANGELOG = DOCS_PATH.joinpath("release-notes.md")
-ROOT_CHANGELOG = REPO_ROOT.joinpath("CHANGELOG.md")
-
-DOCS_CONTRIBUTING = DOCS_PATH.joinpath("contrib", "index.md")
-ROOT_CONTRIBUTING = REPO_ROOT.joinpath("CONTRIBUTING.md")
-
-FILE_COPY_MAPPING = [
-    # Reference document, copied-to location
-    (ROOT_README, DOCS_README),
-    (ROOT_CHANGELOG, DOCS_CHANGELOG),
-    (ROOT_CONTRIBUTING, DOCS_CONTRIBUTING),
-]
-
 AUTOGEN_MESSAGE = [
     "# DO NOT EDIT THIS FILE!",
     "# This file has been autogenerated by `poetry run pymedphys dev propagate`",
@@ -70,78 +48,21 @@ AUTOGEN_MESSAGE = [
 
 
 def propagate_all(args):
-    if args.copies and args.pyproject:
-        raise ValueError("Cannot pass --copies and --pyproject at the same time.")
-
-    if args.copies:
-        run_copies = True
-        run_pyproject = False
-    elif args.pyproject:
-        run_pyproject = True
-        run_copies = False
-    else:
-        run_copies = True
-        run_pyproject = True
-
     if args.update:
         subprocess.check_call("poetry update", shell=True)
 
-    if run_copies:
-        propagate_file_copies_into_library()
+    propagate_version()
+    propagate_extras()
 
-    if run_pyproject:
-        propagate_version()
-        propagate_extras()
-
-        # Propagation of setup.py last as this has the side effect of building
-        # a distribution file. Want to make sure that this distribution
-        # file includes the above propagations in case someone decides to
-        # use it.
-        propagate_lock_requirements_setup_and_hash()
+    # Propagation of setup.py last as this has the side effect of building
+    # a distribution file. Want to make sure that this distribution
+    # file includes the above propagations in case someone decides to
+    # use it.
+    propagate_lock_requirements_and_hash()
 
 
-def propagate_file_copies_into_library():
-    for original_path, target_path in FILE_COPY_MAPPING:
-        _copy_file_with_autogen_message(original_path, target_path)
-
-
-def _copy_file_with_autogen_message(original_path, target_path):
-    if target_path.suffix == ".md":
-        comment_syntax = ("<!-- ", " -->")
-    elif target_path.suffix == ".rst":
-        comment_syntax = ("..\n    ", "")
[…]
