+++ lib/pymedphys/docs/release-notes.md
@@ -1,1402 +1,0 @@
-<!-- DO NOT EDIT THIS FILE! -->
-<!-- This file has been autogenerated by `poetry run pymedphys dev propagate` -->
-<!-- Please instead edit the file found at: -->
-<!--     pymedphys/CHANGELOG.md -->
-<!-- and then run `poetry run pymedphys dev propagate --copies` -->
-
-<!-- markdownlint-disable MD024 MD039 -->
-
-# Release Notes
-
-All notable changes to this project will be documented in this file.
-
-This project adheres to
-[Semantic Versioning](https://semver.org/spec/v2.0.0.html).
-
-## [Unreleased]
-
-### New features and enhancements
-
-- The Gamma tool now utilises the `logging` module. The `quiet` parameter for
-  `pymedphys.gamma` is deprecated but its use is non-breaking (for now).
-
-## [0.38.0]
-
-### News around this release
-
-- [Phillip Chlap](https://github.com/pchlap) has joined as a pymedphys Maintainer
-- Thanks to [Cedric](https://github.com/cedricbelanger) for reporting and fixing DICOM coordinates bug: [Discourse](https://pymedphys.discourse.group/t/gamma-index-calculation-possible-issues/191/2)
-- Privacy related bug within the experimental components of pymedphys (not stable) has been fixed. See bug fix notes.
-
-### Bug Fixes
-
-- Pseudonymise was susceptible to brute-force reversal, mitigated in [Pull Request](https://github.com/pymedphys/pymedphys/pull/1580)
-- DICOM Coordinates
-  - Fix pixel spacing bug [Pull Request](https://github.com/pymedphys/pymedphys/pull/1571)
-- Pinnacle Export Tool
-  - Write DICOM files to be more compatible with other software [Pull Request](https://github.com/pymedphys/pymedphys/pull/1578)
-- Daily Winston-lutz
-  - Small workaround [Pull Request](https://github.com/pymedphys/pymedphys/pull/1565)
-
-### New features and enhancements
-
-- Pinnacle Export Tool
-  - A new CLI option is provided to append the patient MRN to the output path for convenience. Thanks [crewso](https://github.com/crcrewso) for submitting this [Pull Request](https://github.com/pymedphys/pymedphys/pull/1608)
-
-## [0.37.1]
-
-### Bug fixes
-
-- Make it so that Mephysto files that have non-unicode characters can still
-  be opened.
-- Varying penumbra and ball bearing diameter within the Monthly WLutz
-  application now accepts floating point numbers as well as numbers smaller
-  than the default.
-- Fixed a bug where `Delivery.from_monaco` wouldn't be able to load beams
-  where the stop angle of one beam was +180 and the start angle of the
-  subsequent beam was -180 (or visa-versa). The conversion from IEC to bipolar
-  is now handled on a "per-beam" basis. This bug affected the MetersetMap
-  application.
-
-### Non-API changing enhancements
-
-- The daily WLutz application can now have its bb size configured.
-- Site can now be passed as a URL parameter to the daily WLutz application.
-- Improved error messaging around the configuration of the iView machine alias.
-
-## [0.37.0]
-
-### News around this release
-
-- [Jake Rembish](https://github.com/rembishj) has been undergoing his PhD with
-  the output of his work being contributed to PyMedPhys. This release coincides
-  with the culmination of Jake's PhD and this version will be the one
-  referenced within his dissertation. It has been brilliant to see both Jake
-  and his project grow to be what it is today. Thank you Jake 😊.
-
-### Breaking changes
-
-- The configuration for the TRF CLI now utilises the centralised `config.toml`
-  file instead of the prior `CSV` files.
-
-### New features and enhancements
-
-- A range of application have been added to and improved on within the
-  PyMedPhys GUI.
[…]
