### 2022-06-16 412a51d
remove setup.py
