### 2022-06-16 74b8236
remove reference to setup.py
