+++ .github/workflows/library.yml
@@ -4,7 +4,7 @@ on:
   push:
     branches:
       - main
-  pull_request_target:  # https://github.com/Drassil/action-package-version-bump/blob/a4caade1bd6ec37db508f19812355a30506eee9c/.github/workflows/test.yml?rgh-link-date=2022-03-27T14%3A31%3A49Z#L8-L9
+  pull_request_target: # https://github.com/Drassil/action-package-version-bump/blob/a4caade1bd6ec37db508f19812355a30506eee9c/.github/workflows/test.yml?rgh-link-date=2022-03-27T14%3A31%3A49Z#L8-L9
   release:
     types:
       - created
@@ -369,19 +369,11 @@ jobs:
           git config --local user.email "propagate-bot@pymedphys.com"
           git config --local user.name "PyMedPhys Propagate Bot"
 
-          # When on the main branch REPO is not set
-          if [ ! -z "${REPO}" ]
-          then
-            # Designed to handle the case where a fork has been created
-            # Still might need more refinement/testing.
-            git remote set-url origin "https://github.com/${REPO}.git"
-          fi
-
           git fetch origin $BRANCH
           git checkout $BRANCH
 
-          git add poetry.lock setup.py
-          git commit -m "commit changes to poetry.lock and setup.py" || true
+          git add poetry.lock
+          git commit -m "commit changes to poetry.lock" || true
 
           echo "::set-output name=status::committed"
 
@@ -389,7 +381,6 @@ jobs:
           git diff-index --quiet HEAD -- || echo "::set-output name=status::failed"
         env:
           BRANCH: ${{ github.event.pull_request.head.ref }}
-          REPO: ${{ github.event.pull_request.head.repo.full_name }}
 
       - name: Fix any issues
         if: steps.propagate-test.outputs.status == 'failed'
@@ -403,7 +394,7 @@ jobs:
         if: steps.propagate-test.outputs.status == 'failed'
         uses: ad-m/github-push-action@v0.6.0
         with:
-          github_token: ${{ secrets.GITHUB_TOKEN }}
+          github_token: ${{ secrets.PYMEDPHYS_BOT_TOKEN }}
           branch: ${{ github.event.pull_request.head.ref }}
 
       ## Pyright
