+++ .github/workflows/library.yml
@@ -76,7 +76,6 @@ jobs:
             "slow",
             # "binary",  # Skip for now
             "pyright",
-            "propagate",
           ]
 
         exclude:
@@ -92,8 +91,6 @@ jobs:
           - os: "windows"
             task: "pyright"
           - os: "windows"
-            task: "propagate"
-          - os: "windows"
             task: "cypress"
 
           - os: "macos"
@@ -108,8 +105,6 @@ jobs:
             task: "binary"
           - os: "macos"
             task: "pyright"
-          - os: "macos"
-            task: "propagate"
 
         include:
           # Version numbers chosen so that highest and lowest Python version is
@@ -128,13 +123,6 @@ jobs:
 
     steps:
       - uses: actions/checkout@v2
-        if: matrix.task != 'propagate'
-
-      - uses: actions/checkout@v2
-        if: matrix.task == 'propagate'
-        with:
-          persist-credentials: false
-          fetch-depth: 0
 
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v2
@@ -298,10 +286,11 @@ jobs:
 
       ## Install remaining packages for pyright
 
+      # TODO: Likely best to make an explicit pyright extra and delete the propagate extra
       - name: Install a propagate environment
-        if: matrix.task == 'pyright' || matrix.task == 'propagate'
+        if: matrix.task == 'pyright'
         run: |
-          poetry install -E propagate || poetry lock --no-update && poetry install -E propagate
+          poetry install -E propagate
 
       ## Update all packages and then install with tests environment
 
@@ -355,47 +344,6 @@ jobs:
           python test.py
 
       # -----------------------------------------------------------------------------
-
-      # TESTS
-
-      ## Propagate
-      - name: See if propagate is needed
-        if: matrix.task == 'propagate'
-        id: propagate-test
-        run: |
-          echo "::set-output name=status::before"
-          poetry run pymedphys dev propagate
-
-          git config --local user.email "propagate-bot@pymedphys.com"
-          git config --local user.name "PyMedPhys Propagate Bot"
-
-          git fetch origin $BRANCH
-          git checkout $BRANCH
-
-          git add poetry.lock
-          git commit -m "commit changes to poetry.lock" || true
-
-          echo "::set-output name=status::committed"
-
-          git status
-          git diff-index --quiet HEAD -- || echo "::set-output name=status::failed"
-        env:
-          BRANCH: ${{ github.event.pull_request.head.ref }}
-
-      - name: Fix any issues
-        if: steps.propagate-test.outputs.status == 'failed'
-        run: |
-          git config --local user.email "propagate-bot@pymedphys.com"
-          git config --local user.name "PyMedPhys Propagate Bot"
-
-          git commit -m "Fix propagate failures" -a
-
-      - name: Push changes
-        if: steps.propagate-test.outputs.status == 'failed'
-        uses: ad-m/github-push-action@v0.6.0
-        with:
-          github_token: ${{ secrets.PYMEDPHYS_BOT_TOKEN }}
-          branch: ${{ github.event.pull_request.head.ref }}
 
       ## Pyright
 
