### 2022-06-16 1b15ada
add propagate CI back in, but have it be just a test for whether or not is needed, not an automated fix
