+++ setup.py
@@ -1,399 +1,0 @@
-# -*- coding: utf-8 -*-
-
-# DO NOT EDIT THIS FILE!
-# This file has been autogenerated by `poetry run pymedphys dev propagate`
-
-from setuptools import setup
-
-package_dir = {"": "lib"}
-
-packages = [
-    "pymedphys",
-    "pymedphys._base",
-    "pymedphys._data",
-    "pymedphys._dev",
-    "pymedphys._dicom",
-    "pymedphys._dicom.anonymise",
-    "pymedphys._dicom.connect",
-    "pymedphys._dicom.constants",
-    "pymedphys._dicom.ct",
-    "pymedphys._dicom.delivery",
-    "pymedphys._dicom.rtplan",
-    "pymedphys._dicom.structure",
-    "pymedphys._dicom.utilities",
-    "pymedphys._electronfactors",
-    "pymedphys._experimental",
-    "pymedphys._experimental.autosegmentation",
-    "pymedphys._experimental.chartchecks",
-    "pymedphys._experimental.fileformats",
-    "pymedphys._experimental.fileformats.mapcheck",
-    "pymedphys._experimental.fileformats.mephysto",
-    "pymedphys._experimental.fileformats.profiler",
-    "pymedphys._experimental.film",
-    "pymedphys._experimental.paulking",
-    "pymedphys._experimental.pedromartinez",
-    "pymedphys._experimental.pedromartinez.oncentra",
-    "pymedphys._experimental.pedromartinez.utils",
-    "pymedphys._experimental.pinnacle",
-    "pymedphys._experimental.plancomplexity",
-    "pymedphys._experimental.pseudonymisation",
-    "pymedphys._experimental.quickcheck",
-    "pymedphys._experimental.serviceplans",
-    "pymedphys._experimental.streamlit",
-    "pymedphys._experimental.streamlit.apps",
-    "pymedphys._experimental.streamlit.apps.wlutz",
-    "pymedphys._experimental.streamlit.utilities",
-    "pymedphys._experimental.streamlit.utilities.dicom",
-    "pymedphys._experimental.streamlit.utilities.iview",
-    "pymedphys._experimental.vendor.pylinac_vendored",
-    "pymedphys._experimental.vendor.pylinac_vendored.core",
-    "pymedphys._experimental.wlutz",
-    "pymedphys._gamma",
-    "pymedphys._gamma.api",
-    "pymedphys._gamma.implementation",
-    "pymedphys._gamma.utilities",
-    "pymedphys._icom",
-    "pymedphys._imports",
-    "pymedphys._imports.slow",
-    "pymedphys._losslessjpeg",
-    "pymedphys._metersetmap",
-    "pymedphys._metersetmap.delivery",
-    "pymedphys._metersetmap.plt",
-    "pymedphys._mocks",
-    "pymedphys._monaco",
-    "pymedphys._mosaiq",
-    "pymedphys._streamlit",
-    "pymedphys._streamlit.apps",
-    "pymedphys._streamlit.apps.metersetmap",
-    "pymedphys._streamlit.server",
-    "pymedphys._streamlit.utilities",
-    "pymedphys._trf",
-    "pymedphys._trf.decode",
-    "pymedphys._trf.manage",
-    "pymedphys._utilities",
-    "pymedphys._utilities.algorithms",
-    "pymedphys._utilities.constants",
-    "pymedphys._utilities.filehash",
-    "pymedphys._utilities.transforms",
-    "pymedphys._vendor",
-    "pymedphys._vendor.apipkg",
-    "pymedphys.beta",
-    "pymedphys.cli",
-    "pymedphys.cli.experimental",
-    "pymedphys.docs",
[…]
