+++ .readthedocs.yml
@@ -6,8 +6,7 @@ build:
     python: "3.8"
   jobs:
     pre_build:
-      # Generate the Sphinx configuration for this Jupyter Book so it builds.
-      - "jupyter-book config sphinx lib/pymedphys/docs/"
+      - "pymedphys dev docs --prep"
 
 python:
   install:
