+++ _sources/README.rst
@@ -1,14 +1,3 @@
-..
-    DO NOT EDIT THIS FILE!
-..
-    This file has been autogenerated by `poetry run pymedphys dev propagate`
-..
-    Please instead edit the file found at:
-..
-        pymedphys/README.rst
-..
-    and then run `poetry run pymedphys dev propagate --copies`
-
 |logo|
 
 .. |logo| image:: https://github.com/pymedphys/pymedphys/raw/ca501275227f190a77e641a75af925d9070952b6/lib/pymedphys/docs/_static/pymedphys_title.svg
