+++ _sources/contrib/index.md
@@ -1,9 +1,3 @@
-<!-- DO NOT EDIT THIS FILE! -->
-<!-- This file has been autogenerated by `poetry run pymedphys dev propagate` -->
-<!-- Please instead edit the file found at: -->
-<!--     pymedphys/CONTRIBUTING.md -->
-<!-- and then run `poetry run pymedphys dev propagate --copies` -->
-
 # Overview
 
 This document is a work in progress.
