+++ _sources/release-notes.md
@@ -1,9 +1,3 @@
-<!-- DO NOT EDIT THIS FILE! -->
-<!-- This file has been autogenerated by `poetry run pymedphys dev propagate` -->
-<!-- Please instead edit the file found at: -->
-<!--     pymedphys/CHANGELOG.md -->
-<!-- and then run `poetry run pymedphys dev propagate --copies` -->
-
 <!-- markdownlint-disable MD024 MD039 -->
 
 # Release Notes
