### 2022-06-17 pymedphys/pymedphys#1667 @Matthew-Jennings
> I would prefer to abstract it a bit. the library technical reference is oriented towards individual tasks. I would prefer to abstract to these three kinds of tasks:
>
> Data Conversion/extraction from the proprietary formats of a variety of Radiotherapy systems: Two treatment planning systems (I thought I saw Monaco conversion in there somewhere, and Pinnacle is clearly addressed) One Oncology Informatics System (MOSAIQ, but the actual work done is extraction of the meter set map) One Linear accelerator vendor family of systems (log/TRF files as well as iCOM listener)
>
> Manipulation of standard format Radiation Therapy data to address scenarios that are otherwise time consuming and awkward to do with the commercial clinical tools, and for which those commercial tools are not really intended. Anonymisation, extraction of dose information to CSV for plotting, manipulation of contour files to merge them or to adjust scaling of relative electron density, and switching treatment machine names in plans to migrate to a matching machine.
>
> Calculation of the Gamma Index

Fair enough. I think that all sounds great!

>
> And maybe someone can put the above and what I mentioned before together

I _might_ have some time today.
