+++ .readthedocs.yml
@@ -11,8 +11,6 @@ build:
 python:
   install:
     - requirements: requirements-docs.txt
-    - method: pip
-      path: .
 
 sphinx:
   builder: html
