+++ _sources/lib/ref/index.rst
@@ -1,17 +1,0 @@
-=========
-Reference
-=========
-
-PyMedPhys has a range of modules that are able to be exposed from within Python
-itself. Below is the documentation for a range of these libraries.
-
-.. toctree::
-    :maxdepth: 2
-
-    ../../genindex
-    dicom
-    gamma
-    mosaiq
-    metersetmap
-    trf
-    experimental/index
