+++ _sources/lib/tutes/gamma/index.rst
@@ -1,8 +1,0 @@
-=====
-Gamma
-=====
-
-.. toctree::
-    :maxdepth: 2
-
-    gamma_tutorial
