+++ _sources/lib/howto/gamma/index.rst
@@ -1,9 +1,0 @@
-=====
-Gamma
-=====
-
-.. toctree::
-    :maxdepth: 2
-
-    effect-of-noise
-    from-dicom
