+++ _sources/lib/howto/index.md
@@ -1,11 +1,0 @@
-
-# How-to Guides
-
-```{toctree}
-:maxdepth: 2
-
-get-started
-gamma/index
-add-a-linac
-tunnels/index
-```

[…]
