### 2022-06-18 cbad840
Add explanatory comment
Co-authored-by: Simon Biggs <simon.biggs@radiotherapy.ai>
