### 2022-06-21 d982601
dotnet new
