+++ csharp/Class1.cs
@@ -1,0 +1,5 @@
+﻿namespace csharp;
+public class Class1
+{
+
+}
