+++ csharp/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace csharp;
-public class Class1
-{
-
-}
