+++ csharp/ConsoleHelloWorld/Program.cs
@@ -1,3 +1,0 @@
-﻿// See https://aka.ms/new-console-template for more information
-Console.WriteLine("Hello, World!");
-Console.WriteLine("The current time is " + DateTime.Now);
