+++ proto/compile.sh
@@ -1,0 +1,3 @@
+poetry run python -m grpc_tools.protoc \
+  --python_betterproto_out=../lib/pymedphys/_proto/ \
+  -I ../. ../proto/*.proto
