+++ requirements.txt
@@ -44,7 +44,7 @@ ipython==8.4.0; python_version >= "3.8"
 ipywidgets==7.7.0; python_version >= "3.7"
 jedi==0.18.1; python_version >= "3.8"
 jeepney==0.8.0; sys_platform == "linux" and python_version >= "3.7"
-jinja2==3.1.2; python_version >= "3.7"
+jinja2==2.11.3; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
 joblib==1.1.0; python_version >= "3.8"
 jsonschema==3.2.0; python_version >= "3.7"
 jupyter-client==7.3.4; python_full_version >= "3.7.0" and python_version >= "3.7"
@@ -54,13 +54,13 @@ jupyterlab-widgets==1.1.0; python_version >= "3.7"
 keyring==23.6.0; python_version >= "3.7"
 kiwisolver==1.4.2; python_version >= "3.7"
 lxml==4.9.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
-markupsafe==2.1.1; python_version >= "3.7"
+markupsafe==2.0.1; python_version >= "3.6"
 matplotlib-inline==0.1.3; python_version >= "3.8"
 matplotlib==3.5.2; python_version >= "3.7"
 mistune==0.8.4; python_version >= "3.7"
 natsort==8.1.0; python_version >= "3.6"
 nbclient==0.5.13; python_full_version >= "3.7.0" and python_version >= "3.7"
-nbconvert==6.5.0; python_version >= "3.7"
+nbconvert==6.4.5; python_version >= "3.7"
 nbformat==5.4.0; python_full_version >= "3.7.0" and python_version >= "3.7"
 nest-asyncio==1.5.5; python_full_version >= "3.7.0" and python_version >= "3.7"
 networkx==2.8.3; python_version >= "3.8"
@@ -121,10 +121,10 @@ streamlit-ace==0.1.0; python_version >= "3.6"
 streamlit==1.4.0; python_version >= "3.6"
 tenacity==8.0.1; python_version >= "3.6"
 terminado==0.15.0; python_version >= "3.7"
+testpath==0.6.0; python_version >= "3.7"
 threadpoolctl==3.1.0; python_version >= "3.8"
 tifffile==2022.5.4; python_version >= "3.8"
 timeago==1.0.15
-tinycss2==1.1.1; python_version >= "3.7"
 toml==0.10.2; (python_version >= "2.6" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
 tomli==2.0.1; python_version >= "3.7"
 tomlkit==0.11.0; python_version >= "3.6" and python_version < "4.0"

[…]
