+++ lib/pymedphys/_experimental/_proto/pymedphys/__init__.py
@@ -1,5 +1,5 @@
 # Generated by the protocol buffer compiler.  DO NOT EDIT!
-# sources: proto/gamma.proto
+# sources: proto/pymedphys.proto
 # plugin: python-betterproto
 from dataclasses import dataclass
 from typing import Dict, List
@@ -13,9 +13,9 @@ from betterproto.grpc.grpclib_server import ServiceBase
 class GammaRequest(betterproto.Message):
     token: str = betterproto.string_field(1)
     axes_reference: "Double2DArray" = betterproto.message_field(2)
-    dose_reference: "DoubleNdArray" = betterproto.message_field(3)
+    dose_reference: "DoubleArray" = betterproto.message_field(3)
     axes_evaluation: "Double2DArray" = betterproto.message_field(4)
-    dose_evaluation: "DoubleNdArray" = betterproto.message_field(5)
+    dose_evaluation: "DoubleArray" = betterproto.message_field(5)
     dose_percent_threshold: float = betterproto.double_field(6)
     distance_threshold: float = betterproto.double_field(7)
     lower_percent_dose_cutoff: float = betterproto.double_field(8)
@@ -29,7 +29,7 @@ class GammaRequest(betterproto.Message):
 
 @dataclass(eq=False, repr=False)
 class GammaReply(betterproto.Message):
-    data: "DoubleNdArray" = betterproto.message_field(1)
+    data: "DoubleArray" = betterproto.message_field(1)
 
 
 @dataclass(eq=False, repr=False)
@@ -48,10 +48,10 @@ class Double3DArray(betterproto.Message):
 
 
 @dataclass(eq=False, repr=False)
-class DoubleNdArray(betterproto.Message):
-    array1_d: "Double1DArray" = betterproto.message_field(1, group="type")
-    array2_d: "Double2DArray" = betterproto.message_field(2, group="type")
-    array3_d: "Double3DArray" = betterproto.message_field(3, group="type")
+class DoubleArray(betterproto.Message):
+    array_1_d: "Double1DArray" = betterproto.message_field(1, group="type")
+    array_2_d: "Double2DArray" = betterproto.message_field(2, group="type")
+    array_3_d: "Double3DArray" = betterproto.message_field(3, group="type")
 
 
 class GammaServiceStub(betterproto.ServiceStub):
@@ -60,9 +60,9 @@ class GammaServiceStub(betterproto.ServiceStub):
         *,
         token: str = "",
         axes_reference: "Double2DArray" = None,
-        dose_reference: "DoubleNdArray" = None,
+        dose_reference: "DoubleArray" = None,
         axes_evaluation: "Double2DArray" = None,
-        dose_evaluation: "DoubleNdArray" = None,
+        dose_evaluation: "DoubleArray" = None,
         dose_percent_threshold: float = 0.0,
         distance_threshold: float = 0.0,
         lower_percent_dose_cutoff: float = 0.0,
@@ -94,7 +94,9 @@ class GammaServiceStub(betterproto.ServiceStub):
         request.random_subset = random_subset
         request.ram_available = ram_available
 
-        return await self._unary_unary("/gamma.GammaService/Gamma", request, GammaReply)
+        return await self._unary_unary(
+            "/pymedphys.GammaService/Gamma", request, GammaReply
+        )
 
 
 class GammaServiceBase(ServiceBase):
@@ -102,9 +104,9 @@ class GammaServiceBase(ServiceBase):
         self,
         token: str,
         axes_reference: "Double2DArray",
-        dose_reference: "DoubleNdArray",
+        dose_reference: "DoubleArray",
         axes_evaluation: "Double2DArray",
-        dose_evaluation: "DoubleNdArray",
[…]

[…]
