+++ lib/pymedphys/_experimental/_proto/pymedphys/__init__.py
@@ -1,153 +1,0 @@
-# Generated by the protocol buffer compiler.  DO NOT EDIT!
-# sources: proto/pymedphys.proto
-# plugin: python-betterproto
-from dataclasses import dataclass
-from typing import Dict, List
-
-import betterproto
-import grpclib
-from betterproto.grpc.grpclib_server import ServiceBase
-
-
-@dataclass(eq=False, repr=False)
-class GammaRequest(betterproto.Message):
-    token: str = betterproto.string_field(1)
-    axes_reference: "Double2DArray" = betterproto.message_field(2)
-    dose_reference: "DoubleArray" = betterproto.message_field(3)
-    axes_evaluation: "Double2DArray" = betterproto.message_field(4)
-    dose_evaluation: "DoubleArray" = betterproto.message_field(5)
-    dose_percent_threshold: float = betterproto.double_field(6)
-    distance_threshold: float = betterproto.double_field(7)
-    lower_percent_dose_cutoff: float = betterproto.double_field(8)
-    interpolation_fraction: float = betterproto.double_field(9)
-    max_gamma: float = betterproto.double_field(10)
-    local_gamma: bool = betterproto.bool_field(11)
-    global_normalisation: float = betterproto.double_field(12)
-    random_subset: int = betterproto.int64_field(13)
-    ram_available: int = betterproto.int64_field(14)
-
-
-@dataclass(eq=False, repr=False)
-class GammaReply(betterproto.Message):
-    data: "DoubleArray" = betterproto.message_field(1)
-
-
-@dataclass(eq=False, repr=False)
-class Double1DArray(betterproto.Message):
-    data: List[float] = betterproto.double_field(1)
-
-
-@dataclass(eq=False, repr=False)
-class Double2DArray(betterproto.Message):
-    data: List["Double1DArray"] = betterproto.message_field(1)
-
-
-@dataclass(eq=False, repr=False)
-class Double3DArray(betterproto.Message):
-    data: List["Double2DArray"] = betterproto.message_field(1)
-
-
-@dataclass(eq=False, repr=False)
-class DoubleArray(betterproto.Message):
-    array_1_d: "Double1DArray" = betterproto.message_field(1, group="type")
-    array_2_d: "Double2DArray" = betterproto.message_field(2, group="type")
-    array_3_d: "Double3DArray" = betterproto.message_field(3, group="type")
-
-
-class GammaServiceStub(betterproto.ServiceStub):
-    async def gamma(
-        self,
-        *,
-        token: str = "",
-        axes_reference: "Double2DArray" = None,
-        dose_reference: "DoubleArray" = None,
-        axes_evaluation: "Double2DArray" = None,
-        dose_evaluation: "DoubleArray" = None,
-        dose_percent_threshold: float = 0.0,
-        distance_threshold: float = 0.0,
-        lower_percent_dose_cutoff: float = 0.0,
-        interpolation_fraction: float = 0.0,
-        max_gamma: float = 0.0,
-        local_gamma: bool = False,
-        global_normalisation: float = 0.0,
-        random_subset: int = 0,
-        ram_available: int = 0
-    ) -> "GammaReply":
-
-        request = GammaRequest()
-        request.token = token
-        if axes_reference is not None:
-            request.axes_reference = axes_reference
-        if dose_reference is not None:
-            request.dose_reference = dose_reference
-        if axes_evaluation is not None:
-            request.axes_evaluation = axes_evaluation
-        if dose_evaluation is not None:
-            request.dose_evaluation = dose_evaluation
-        request.dose_percent_threshold = dose_percent_threshold
-        request.distance_threshold = distance_threshold
-        request.lower_percent_dose_cutoff = lower_percent_dose_cutoff
-        request.interpolation_fraction = interpolation_fraction
-        request.max_gamma = max_gamma
[…]
