+++ proto/compile.sh
@@ -1,3 +1,0 @@
-poetry run python -m grpc_tools.protoc \
-  --python_betterproto_out=../lib/pymedphys/_experimental/_proto/ \
-  -I ../. ../proto/*.proto
