### 2022-06-22 #4470 @Kevin-Mattheus-Moerman
@SimonBiggs @Matthew-Jennings can you address these comments? :point_up: i.e. clarify software application in the readme and paper?
