### 2022-06-22 #4470 @SimonBiggs
Yup, thanks @Kevin-Mattheus-Moerman.

@sjswerdloff and @Matthew-Jennings have begun the discussion over at:

https://github.com/pymedphys/pymedphys/issues/1667
