### 2022-06-23 54afbaa
Update paper.md
Proposed update to paper.md to include additional details of the applications/tools as requested by the paper reviewers and outlined by @sjswerdloff and @Matthew-Jennings at https://github.com/pymedphys/pymedphys/issues/1667.

If the changes are too wordy or add to much volume I think it would be reasonable to trim the description of tools that make use of data extraction/conversion. I.e. lines 110 to 117 could be cut.
