### 2022-06-23 #4470 @SimonBiggs
@Kevin-Mattheus-Moerman, does the following proposed changes to the paper appropriately address the concerns raised?

https://github.com/pymedphys/pymedphys/suites/7052584802/artifacts/277878183
