+++ requirements.txt
@@ -77,7 +77,7 @@ plotly==5.8.1; python_version >= "3.6"
 premailer==3.10.0
 prometheus-client==0.14.1; python_version >= "3.7"
 prompt-toolkit==3.0.29; python_full_version >= "3.6.2" and python_version >= "3.8"
-protobuf==3.18.1; python_version >= "3.5"
+protobuf==3.18.1; python_version >= "3.6"
 psutil==5.9.1; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
 ptyprocess==0.7.0; sys_platform != "win32" and python_version >= "3.8" and os_name != "nt"
 pure-eval==0.2.2; python_version >= "3.8"
@@ -87,8 +87,8 @@ pycparser==2.21; python_version >= "3.7" and python_full_version < "3.0.0" and s
 pydeck==0.7.1; python_version >= "3.7"
 pydicom==2.3.0; python_full_version >= "3.6.1"
 pygments==2.12.0; python_version >= "3.8"
-pylibjpeg-libjpeg==1.3.1; python_version >= "3.7"
 pylinac==2.3.2
+pymedphys-pylibjpeg-libjpeg==1.3.1; python_version >= "3.7"
 pympler==1.0.1; python_version >= "3.6"
 pymssql==2.2.5
 pynetdicom==2.0.2; python_version >= "3.7"
