+++ pyproject.toml
@@ -70,7 +70,9 @@ fsspec = { version = "*", optional = true }            # groups = ["user", "dev"
 dicompyler-core = { version = "*", optional = true }   # groups = ["user", "dev"]
 scikit-learn = { version = "*", optional = true }      # groups = ["user", "dev", "mosaiq"]
 imageio = { version = "*", optional = true }           # groups = ["user", "dev"]
-pylibjpeg-libjpeg = { version = "*", optional = true } # groups = ["user", "dev"]
+
+streamlit = { version = "*", optional = true }     # groups = ["user", "dev"]
+streamlit-ace = { version = "*", optional = true } # groups = ["user", "dev"]
 
 # NumPy's lower bound here is for typing support
 numpy = { version = ">=1.20.2", optional = true } # groups = ["user", "dev", "icom", "docs"]
@@ -85,11 +87,11 @@ pydicom = { version = ">=2.0.0", optional = true } # groups = ["user", "dev", "d
 shapely = { version = ">=1.7.0", optional = true }       # groups = ["user", "dev"]
 scikit-image = { version = ">=0.18.1", optional = true } # groups = ["user", "dev"]
 
-# Streamlit's pinning is due to its current beta nature and the fact that some
-# private APIs are utilised.
-protobuf = { version = "<3.19.0", optional = true }    # groups = ["user", "dev"]
-streamlit = { version = "1.4.0", optional = true }     # groups = ["user", "dev"]
-streamlit-ace = { version = "0.1.0", optional = true } # groups = ["user", "dev"]
+# Utilising the pymedphys fork of pylibjpeg-libjpeg to fix the following issue:
+# https://github.com/pymedphys/pymedphys/issues/1556#issuecomment-1152821285
+pymedphys-pylibjpeg-libjpeg = { version = "1.3.1", optional = true } # groups = ["user", "dev"]
+# pylibjpeg-libjpeg = { version = "*", optional = true } # groups = ["user", "dev"]
+
 
 # PyLinac is pinned due to historical algorithm changes with both minor and
 # patch version bumps.
@@ -182,12 +184,11 @@ dev = [
     "pandas",
     "plotly",
     "pre-commit",
-    "protobuf",
     "psutil",
     "pydicom",
-    "pylibjpeg-libjpeg",
     "pylinac",
     "pylint",
+    "pymedphys-pylibjpeg-libjpeg",
     "pymssql",
     "pynetdicom",
     "pyoxidizer",
@@ -263,10 +264,9 @@ user = [
     "packaging",
     "pandas",
     "plotly",
-    "protobuf",
     "pydicom",
-    "pylibjpeg-libjpeg",
     "pylinac",
+    "pymedphys-pylibjpeg-libjpeg",
     "pymssql",
     "pynetdicom",
     "python-dateutil",
