+++ requirements.txt
@@ -3,22 +3,21 @@ appnope==0.1.3; sys_platform == "darwin" and python_version >= "3.8" and platfor
 argon2-cffi-bindings==21.2.0; python_version >= "3.7"
 argon2-cffi==21.3.0; python_version >= "3.7"
 argue==0.3.1
-astor==0.8.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
 asttokens==2.0.5; python_version >= "3.8"
 attrs==21.4.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
 backcall==0.2.0; python_version >= "3.8"
 backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.6" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
-base58==2.1.1; python_version >= "3.6"
 beautifulsoup4==4.11.1; python_full_version >= "3.6.0" and python_version >= "3.7"
 bleach==5.0.0; python_version >= "3.7"
 blinker==1.4; python_version >= "3.6"
 cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0"
-certifi==2022.5.18.1; python_version >= "3.7" and python_version < "4"
+certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
 cffi==1.15.0; sys_platform == "linux" and python_version >= "3.7" and implementation_name == "pypy"
 charset-normalizer==2.0.12; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.5.0"
 click==8.1.3; python_version >= "3.7"
-colorama==0.4.4; python_version >= "3.8" and python_full_version < "3.0.0" and platform_system == "Windows" and sys_platform == "win32" or python_full_version >= "3.5.0" and platform_system == "Windows" and sys_platform == "win32" and python_version >= "3.8"
-cryptography==37.0.2; sys_platform == "linux" and python_version >= "3.7"
+colorama==0.4.5; python_version >= "3.8" and python_full_version < "3.0.0" and platform_system == "Windows" and sys_platform == "win32" or python_full_version >= "3.5.0" and platform_system == "Windows" and sys_platform == "win32" and python_version >= "3.8"
+commonmark==0.9.1; python_full_version >= "3.6.3" and python_full_version < "4.0.0" and python_version >= "3.6"
+cryptography==37.0.3; sys_platform == "linux" and python_version >= "3.7"
 cssselect==1.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
 cssutils==2.4.2; python_version >= "3.7"
 cycler==0.11.0; python_version >= "3.7"
@@ -38,10 +37,10 @@ greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (pl
 idna==3.3; python_version >= "3.7" and python_version < "4"
 imageio==2.19.3; python_version >= "3.7"
 importlib-metadata==4.11.4; python_version < "3.10" and python_version >= "3.7"
-ipykernel==6.13.1; python_version >= "3.7"
+ipykernel==6.15.0; python_version >= "3.7"
 ipython-genutils==0.2.0; python_version >= "3.7"
 ipython==8.4.0; python_version >= "3.8"
-ipywidgets==7.7.0; python_version >= "3.7"
[…]
