+++ requirements-icom.txt
@@ -1,2 +1,2 @@
-numpy==1.22.4; python_version >= "3.8"
+numpy==1.23.0; python_version >= "3.8"
 typing-extensions==4.2.0; python_version >= "3.7"
