+++ js/app/src/window.ts
@@ -1,0 +1,109 @@
+// The MIT License (MIT)
+
+// Copyright (c) 2015-2021 Jakub Szwacz
+
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to
+// deal in the Software without restriction, including without limitation the
+// rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+// sell copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+// IN THE SOFTWARE.
+
+// Adapted from
+// https://github.com/szwacz/electron-boilerplate/blob/20aff6cd/src/helpers/window.js
+
+import { app, BrowserWindow, screen } from "electron";
+import jetpack from "fs-jetpack";
+
+const window: (
+  name: string,
+  options: Electron.BrowserWindowConstructorOptions
+) => BrowserWindow = (name, options) => {
+  const userDataDir = jetpack.cwd(app.getPath("userData"));
+  const stateStoreFile = `window-state-${name}.json`;
+  const defaultSize = {
+    width: options.width,
+    height: options.height,
+  };
+  let state = {};
+  let win: BrowserWindow;
+
+  const restore = () => {
+    let restoredState = {};
+    try {
+      restoredState = userDataDir.read(stateStoreFile, "json");
+    } catch (err) {
+      // For some reason json can't be read (might be corrupted).
+      // No worries, we have defaults.
+    }
+    return Object.assign({}, defaultSize, restoredState);
+  };
+
+  const getCurrentPosition = () => {
+    const position = win.getPosition();
+    const size = win.getSize();
+    return {
+      x: position[0],
+      y: position[1],
+      width: size[0],
+      height: size[1],
+    };
+  };
+
+  const windowWithinBounds = (windowState, bounds) => {
+    return (
+      windowState.x >= bounds.x &&
+      windowState.y >= bounds.y &&
+      windowState.x + windowState.width <= bounds.x + bounds.width &&
+      windowState.y + windowState.height <= bounds.y + bounds.height
+    );
+  };
+
+  const resetToDefaults = () => {
+    const bounds = screen.getPrimaryDisplay().bounds;
+    return Object.assign({}, defaultSize, {
+      x: (bounds.width - defaultSize.width) / 2,
+      y: (bounds.height - defaultSize.height) / 2,
+    });
+  };
+
+  const ensureVisibleOnSomeDisplay = (windowState) => {
+    const visible = screen.getAllDisplays().some((display) => {
+      return windowWithinBounds(windowState, display.bounds);
+    });
+    if (!visible) {
+      // Window is partially or fully not visible now.
+      // Reset it to safe defaults.
+      return resetToDefaults();
+    }
+    return windowState;
[…]
