+++ requirements.txt
@@ -66,7 +66,7 @@ networkx==2.8.4; python_version >= "3.8"
 notebook==6.4.12; python_version >= "3.7"
 numpy==1.23.0; python_version >= "3.8"
 packaging==21.3; python_version >= "3.6"
-pandas==1.4.2; python_version >= "3.8"
+pandas==1.4.3; python_version >= "3.8"
 pandocfilters==1.5.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
 parso==0.8.3; python_version >= "3.8"
 pexpect==4.8.0; sys_platform != "win32" and python_version >= "3.8"
@@ -116,7 +116,7 @@ shapely==1.8.2; python_version >= "3.6"
 six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
 smmap==5.0.0; python_version >= "3.7"
 soupsieve==2.3.2.post1; python_full_version >= "3.6.0" and python_version >= "3.7"
-sqlalchemy==1.4.37; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
+sqlalchemy==1.4.38; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
 stack-data==0.3.0; python_version >= "3.8"
 streamlit-ace==0.1.1; python_version >= "3.6"
 streamlit==1.10.0; python_version >= "3.6"
