### 2022-06-24 fa59f16
create mult-platform binary build CI
