+++ lib/pymedphys/_dev/paths.py
@@ -1,0 +1,19 @@
+# Copyright (C) 2022 Simon Biggs
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+import pathlib
+
+REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent.parent.parent
+LIBRARY_PATH = REPO_ROOT.joinpath("lib", "pymedphys")
