### 2022-06-24 7eb3c57
make sure the lowest python version is actually utilised in CI
