+++ .github/workflows/library.yml
@@ -109,11 +109,9 @@ jobs:
 
         include:
           # Version numbers chosen so that highest and lowest Python version is
-          # utilised, 3.7 is chosen on Windows to align with Jupyter Book until
-          # https://github.com/executablebooks/jupyter-book/issues/906 is fixed.
-          # See https://jupyterbook.org/advanced/advanced.html#working-on-windows
+          # utilised
           - os: "ubuntu"
-            python-version: "3.10"
+            python-version: "3.7"
           - os: "windows"
             python-version: "3.10"
           - os: "macos"
