### 2022-06-24 53a662d
ignore filenotfound errors in the binary build
