### 2022-06-24 366f591
fix ubuntu python version in CI
