### 2022-06-24 32d528c
add os to the binary name
