+++ .github/workflows/library.yml
@@ -93,8 +93,6 @@ jobs:
             task: "propagate"
           - os: "windows"
             task: "cypress"
-          - os: "windows"
-            task: "build"
 
           - os: "macos"
             task: "stackoverflow"
