### 2022-06-25 #4470 @Kevin-Mattheus-Moerman
@osorensen do you think you can help edit this Python submission relating to medical physics? I thought of you given your links to medical (neuro)imaging, however it may be somewhat off topic. If you cannot take this one, then I can take it but I would prefer if you could since I have quite a full plate at the moment.
