### 2022-06-25 #4470 @ProfLeao
> would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:

@osorensen, of course i can. It will be a pleasure
