### 2022-06-26 #4470 @SimonBiggs
> @SimonBiggs the readme appears unchanged, is that correct? On the paper
>
> The summary reads:
>
> > PyMedPhys is an open-source medical physics library built for Python by a diverse community that values and prioritizes code sharing, review, continuous improvement, and peer development.
>
> This seems to simply say, this is open source software and we follow the open source philosophy. This is of course standard and a requirement of JOSS. However it does not say much about what the software is for. You can consider shortening this if you like.
>
> > PyMedPhys aims to simplify and enhance both research and clinical work related to medical physics. It is inspired by the Astropy Project (Astropy Collaboration, 2013); a highly successful collaborative work of our physics peers in astronomy.
>
> I read this as: PyMedPhys enhances research/clinical work in "medical physics". Medical physics can be anything at this point (MRI, radiation, simulation of cardiac ablation therapy,...) so the reader is unsure. The next sentence then points to Astropy, and it says it is a "highly successful collaborative work of our physics peers in astronomy.". After looking at the paper a bit more I think I now understand that you were inspired by Astropy in the sense that it seems to be a collection of methods and algorithms used in astrophysics. And Astropy aspires to be standard/canonical open source and library used by astrophysicists. So anything useful to astrophysicists will be contained in this one open source ecosystem. So it looks like you want PyMedPhys to be like Astropy in that way but to instead be an open source library that collects all matter of methods and algorithms relevant to the medical physics community. And medical physics is kept broad/general there and includes radiation work. I think I get it now but perhaps your text can be improved. This is only a recommendation but perhaps this would help (please feel completely free to reject this recommendation or change it completely): `PyMedPhys aims to simplify and enhance both research and clinical work related to medical physics. It is inspired by the Astropy Project (Astropy Collaboration, 2013); which forms a common core package for Astronomy in Python. Similarly PyMedPhys aspires to be a canonical open source library, but instead for medical physics. PyMedPhys currently includes [insert some functionality description], and aims to cover X, Y, Z and other medical physics software utilities in future work.` Something like the above clarifies the goal of the package and the functionality more clearly. Again, just a recommendation.

@JacobMcAloney, I really like the suggestion above. Would you up for another adjustment taking into account that recommendation?
