### 2022-06-27 #4470 @Matthew-Jennings
Thank you, @Kevin-Mattheus-Moerman, for your helpful initial review.

> @SimonBiggs the readme appears unchanged, is that correct?

So far, yes, though we do see the merit of fleshing it out as per @danasolav's feedback. Would the README in its current form preclude publication in JOSS? If so, what specifically would need to be included from JOSS' perspective? One point that comes to my mind is that, unlike a JOSS paper summary, the README's missing content probably ought to be (primarily) targeted at a _specialist_ audience.

I'd say that we're willing to cater to your/the reviewers' preferences here, just noting that it might take some time (perhaps a week or two) for those edits to be reflected in the README.
