### 2022-06-27 0b746fa
continue to flesh out, deep dive on exe doc
