### 2022-06-27 041a28c
last sections of how-to binary docs
