+++ lib/pymedphys/docs/contrib/dive/create-streamlit-exe.md
@@ -15,9 +15,9 @@ All code links will be pointing to the code base at that commit hash.
 To actually build and create the binary all you actually need to do is the following:
 
 - Install [Poetry](https://python-poetry.org/docs/#installation)
-- Install [Node](https://nodejs.org/en/) and [Yarn](https://classic.yarnpkg.com/en/docs/install#debian-stable)
+- Install [Node](https://nodejs.org/en/) and [Yarn 1.x](https://classic.yarnpkg.com/en/docs/install#debian-stable)
 - Install PyOxidizer by installing all the project build dependencies:
-  - poetry install -E build
+  - `poetry install -E build -E cli`
 - Then run `poetry run pymedphys dev build --install`
   - Which runs [this](https://github.com/pymedphys/pymedphys/blob/836f272d092f294099bb51db05bab80d2bfcb628/lib/pymedphys/_dev/build.py#L28-L51) under the hood.
 
@@ -43,7 +43,7 @@ given below.
 
 ## Creating a self contained PyMedPhys python bundle
 
-### Justification of approach
+### Justification
 
 To create the self contained Python environment PyOxidizer was utilised. This
 tool contains a range of embeddable Python distributions for each OS. There are
@@ -87,12 +87,164 @@ resource location
 This is due to Streamlit using [that file directory location as an indicator](https://github.com/streamlit/streamlit/blob/953dfdbeb51a4d0cb4ddb81aaad8e4321fe5db73/lib/streamlit/config.py#L255-L267)
 that Streamlit is not running in developer mode.
 
+### Running the build
+
 Once that configuration file is created, and once PyOxidizer is installed then
 running `poetry run pyoxidizer build install` will create a `pymedphys` binary
 within a `dist` directory within the repository root.
 
-## Setting up electron
+## The use and setting up Electron
+
+### Justification
+
+Now that we have a self contained PyMedPhys distribution, arguably we could be
+done. Also, arguably, bringing in the heavy beast of Electron potentially
+complicates things significantly. However, it does have a few benefits. The
+primary benefits being that electron builder manages the creation of the
+installer as well as management of auto-update capacity, if we want to support
+that in the future.
+
+PyOxidizer does also manage the creation of the installer. However, with my
+trialling of it in the past this installer was only able to run as
+administrator on Windows, which is likely a deal breaker in many scenarios.
+
+Instead of using Electron we could manually handle the installer ourselves. But
+then we'd have to do that independently for each OS. By using electron, we can
+run the same set up on one OS, and have it automatically create installers on
+each target OS with little to no adjustments needed. Having a bulky executable
+created at the other side is a reasonable trade off in my opinion for
+significantly reduced maintenance burden on PyMedPhys.
+
+### Setting up Electron
+
+To set up the Electron code base I began with the following boilerplate code:
+
+> <https://github.com/szwacz/electron-boilerplate>
+
+I significantly stripped it back so as to include as little as possible while
+still having success.
+
+The resulting Electron application code utilised can be found at:
+
+> <https://github.com/pymedphys/pymedphys/tree/836f272d092f294099bb51db05bab80d2bfcb628/js/app>
+
+To install all of the required dependencies run `yarn install` within the
+`js/app` directory. You will need to have both [Node](https://nodejs.org/en/)
[…]
