### 2022-06-27 a12dc83
inline some code snippets
