### 2022-06-30 #4470 @osorensen
>

@Matthew-Jennings, no, the README in its current form does not preclude publication in JOSS. I am looking for reviewers now, and once I find three reviewers, these will have a look at the README and everything else.
