### 2022-06-30 #4470 @osorensen
:wave: @eteq, would you be willing to review this submission for the Journal of Open Source Software? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I realize might this seems far from your field of work, however, this is a medical physics tool which is directly inspired by the Astropy Project. Given your key role in Astropy, if you would be able to do a review with a focus on the software rather than the particular applications, it would be much appreciated.
