### 2022-07-04 addb293
Fix notebook metadata
Remove unused lines in workflows/library due to merging
