### 2022-07-05 #4470 @osorensen
:wave: @brechmos @grlee77 @7omasz would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
