### 2022-07-08 #4555 @ProfLeao
> Hi @ProfLea, yes, you're welcome to add suggestions to the text here. Alternatively you can also open an issue in the source repository. The latter would be easier if you have a larger number of suggested changes.

ok thanks.
