### 2022-07-08 #4555 @SimonBiggs
Hi @ProfLeao,

I don't we need to edit that guidance component of the original license file. Instead, we just make sure to include that boilerplate at the top of each source file:

https://github.com/pymedphys/pymedphys/blob/main/lib/pymedphys/__main__.py#L1-L13
