### 2022-07-11 #4555 @ProfLeao
> Hi @ProfLeao,
>
> Thanks for jumping in to review PyMedPhys 🙂
>
> I don't believe we need to edit that guidance component of the original license file. Instead, we just make sure to include that boilerplate at the top of each source file:
>
> https://github.com/pymedphys/pymedphys/blob/main/lib/pymedphys/__main__.py#L1-L13

Hello @SimonBiggs .  Thanks for the clarification.  Today I will continue to study the software, I will let you know if any other questions arise.
