### 2022-07-12 #4555 @ProfLeao
Hello @SimonBiggs . I noticed that some features were migrated in the directory tree and the examples still do not include some of these changes. With the expansion of software visibility that JOSS can offer in mind, it would be good to update this.
Take, for example, the case of `pinnacle`.

old:
``` Python
from pymedphys.labs.pinnacle import PinnacleExport
```

Now:
``` Python
from pymedphys._experimental.pinnacle import PinnacleExport
```
