### 2022-07-19 pymedphys/pymedphys#1697 @SimonBiggs
JOSS Review Items
https://github.com/openjournals/joss-reviews/issues/4555#issuecomment-1182529100

@crcrewso, might you be open to fixing up this review feedback item from the JOSS paper review?
