### 2022-07-20 pymedphys/pymedphys#1697 @crcrewso
Absolutely. I can do that this week.
