### 2022-07-20 #4555 @gbaltz
Hi,
Looks like several of the links in the documentation section of the readme are broken, namely: "getting started", "library users guide", "CLI users guide", and "General".

Additionally I think the README file is missing a couple of the required sections as part of my Documentation review checklist. Currently missing sections are "Statement of Need" and "Automated Tests". I also think an "Examples" section is maybe missing as it takes a couple clicks digging in the documentation to find examples.
