### 2022-07-20 pymedphys/pymedphys#1697 @SimonBiggs
Beautiful :) thanks @crcrewso
