### 2022-07-20 #4555 @SimonBiggs
> Looks like several of the links in the documentation section of the readme are broken, namely: "getting started", "library users guide", "CLI users guide", and "General".

Hi @Matthew-Jennings, would you be able to take this one?
