### 2022-07-20 #4555 @SimonBiggs
> Additionally I think the README file is missing a couple of the required sections as part of my Documentation review checklist. Currently missing sections are "Statement of Need" and "Automated Tests". I also think an "Examples" section is maybe missing as it takes a couple clicks digging in the documentation to find examples.

@nlambriICH might you be able to take this one?
