### 2022-07-21 pymedphys/pymedphys#1697 @guanfada
@SimonBiggs I have a co-authored paper accepted by PMB today, so I guess tomorrow or later I can get the DOI of this accepted paper. Then we can put it into our JOSS paper as a reference. What do you think?
