### 2022-07-21 #4555 @Matthew-Jennings
> > Looks like several of the links in the documentation section of the readme are broken, namely: "getting started", "library users guide", "CLI users guide", and "General".
>
> Hi @Matthew-Jennings, would you be able to take this one?

@SimonBiggs: Yep, will try to tonight.
