### 2022-07-22 #4555 @nlambriICH
> Currently missing sections are "Statement of Need" and "Automated Tests". I also think an "Examples" section is maybe missing as it takes a couple clicks digging in the documentation to find examples.

@guanfada Can you start adding a draft for these required sections in the README?. We can then work together for the final version.
