### 2022-07-23 479cf3f
reword documentation section of readme. fix links
