+++ paper.bib
@@ -157,3 +157,13 @@
   author={Bert, Christoph and Bertholet, Jenny and Essers, Marion and Heijmen, Ben and Jornet, Nuria and Koutsouveli, Efi and Schwarz, Marco and Bodale, Marin and Casares-Magaz, Oscar and Gerskevitch, Eduard and others},
   url={https://www.efomp.org/uploads/595e3c8a-52d9-440f-b50b-183c3a00cb00/Radiotherapy_cc_2022.pdf}
 }
+
+@article{yang2022PMB,
+  title={Adaptation and dosimetric commissioning of a synchrotron-based proton beamline for FLASH experiments},
+  author={Yang, Ming and Wang, Xiaochun and Guan, Fada and Titt, Uwe and Iga, Kiminori and Jiang, Dadi and Takaoka, Takeshi and Totake, Satoshi and Katayose, Tadashi and Umezawa, Masumi and Schüler, Emil and Frank, Steven J and Lin, Steven Hsesheng and Sahoo, Narayan and Koong, Albert C and Mohan, Radhe and Zhu, Xiaorong Ronald},
+  journal={Physics in Medicine and Biology},
+  year={2022},
+  publisher={IOP Publishing},
+  url={https://iopscience.iop.org/article/10.1088/1361-6560/ac8269/meta}
+  doi={10.1088/1361-6560/ac8269}
+}
