### 2022-07-23 f9c359f
Update README.rst
Added "examples"
Added "statement of need"
I am not familiar with the "Automated Tests"
