+++ paper.bib
@@ -37,6 +37,7 @@
   title={Automating Medical Physics Quality Assurance Tasks in Radiation Oncology},
   author={Rembish, Jacob},
   year={2021},
+  url={https://www.proquest.com/docview/2564568968}
   school={The University of Texas Health Science Center at San Antonio}
 }
 
@@ -91,6 +92,7 @@
   title={MRI only radiotherapy planning using the transfer function estimation algorithm},
   author={Cronholm, Rickard O and Karlsson, Anders and Siversson, Carl},
   year={2020}
+  url={http://www.spectronic.se/files/Whitepaper_TFE_202106.pdf}
 }
 
 @article{milan2019evaluation,
