### 2022-07-24 cc91388
Update README.rst
added "statement of need" section here
