### 2022-07-24 820ab5f
Update paper.bib
Co-authored-by: Matthew Jennings <Centrus.007@gmail.com>
