+++ paper.bib
@@ -37,7 +37,7 @@
   title={Automating Medical Physics Quality Assurance Tasks in Radiation Oncology},
   author={Rembish, Jacob},
   year={2021},
-  url={https://www.proquest.com/docview/2564568968}
+  url={https://www.proquest.com/docview/2564568968},
   school={The University of Texas Health Science Center at San Antonio}
 }
 
