### 2022-07-24 8a152f5
Update paper.bib
Co-authored-by: Matthew Jennings <Centrus.007@gmail.com>
