+++ paper.bib
@@ -91,7 +91,7 @@
 @article{cronholm2020mri,
   title={MRI only radiotherapy planning using the transfer function estimation algorithm},
   author={Cronholm, Rickard O and Karlsson, Anders and Siversson, Carl},
-  year={2020}
+  year={2020},
   url={http://www.spectronic.se/files/Whitepaper_TFE_202106.pdf}
 }
 
