+++ paper.bib
@@ -51,7 +51,7 @@
   year={2020},
   publisher={IOP Publishing},
   doi={10.1088/1361-6560/abb71b},
-	url={https://doi.org/10.1088/1361-6560/abb71b},
+  url={https://doi.org/10.1088/1361-6560/abb71b},
 }
 
 @article{baltz2021validation,
@@ -134,7 +134,7 @@
   publisher={Wiley Online Library},
   journal={Journal of Applied Clinical Medical Physics},
   pages={e13556},
-  year = {2022},
+  year={2022},
   keywords={Monte Carlo independent dose calculation, virtual source model, Varian Linac},
   doi={10.1002/acm2.13556},
   url={https://aapm.onlinelibrary.wiley.com/doi/abs/10.1002/acm2.13556},
