+++ paper.bib
@@ -166,6 +166,6 @@
   journal={Physics in Medicine and Biology},
   year={2022},
   publisher={IOP Publishing},
-  url={https://iopscience.iop.org/article/10.1088/1361-6560/ac8269/meta}
+  url={https://iopscience.iop.org/article/10.1088/1361-6560/ac8269/meta},
   doi={10.1088/1361-6560/ac8269}
 }
