### 2022-07-25 pymedphys/pymedphys#1697 @crcrewso
It appears there are a few JOSS related tasks on my todo list. Is it possible to add the cc-pinnacle-joss-patch branch as a 'version' in the readthedocs configuration?
