### 2022-07-25 pymedphys/pymedphys#1697 @SimonBiggs
> It appears there are a few JOSS related tasks on my todo list. Is it possible to add the cc-pinnacle-joss-patch branch as a 'version' in the readthedocs configuration?

I actually just added you as a maintainer on the readthedocs PyMedPhys project. Feel free to make any branch you choose / wish to be set to a hidden version as you desire :slightly_smiling_face:
