### 2022-07-26 cc6cd13
Fixes #1696 and adds @crcrewso to contributor list
