### 2022-08-03 b2eb107
Add activity schedule by location Mosaiq helper
