### 2022-08-04 d4476e4
Delete warning-pinnacle.rst
