+++ requirements.txt
@@ -56,7 +56,7 @@ lxml==4.9.1; python_version >= "2.7" and python_full_version < "3.0.0" or python
 markupsafe==2.1.1; python_version >= "3.7"
 matplotlib-inline==0.1.3; python_version >= "3.8"
 matplotlib==3.5.2; python_version >= "3.7"
-mistune==0.8.4; python_version >= "3.7"
+mistune==2.0.3; python_version >= "3.7"
 natsort==8.1.0; python_version >= "3.6"
 nbclient==0.5.13; python_full_version >= "3.7.0" and python_version >= "3.7"
 nbconvert==6.5.0; python_version >= "3.7"
