+++ requirements-docs.txt
@@ -54,7 +54,7 @@ markupsafe==2.1.1; python_version >= "3.7" and python_full_version < "3.0.0" or 
 matplotlib-inline==0.1.3; python_version >= "3.8"
 matplotlib==3.5.2; python_version >= "3.7"
 mdit-py-plugins==0.2.8; python_version >= "3.7" and python_version < "4.0"
-mistune==0.8.4; python_version >= "3.7"
+mistune==2.0.3; python_version >= "3.7"
 myst-nb==0.13.2; python_version >= "3.7"
 myst-parser==0.15.2; python_version >= "3.7"
 nbclient==0.5.13; python_full_version >= "3.7.0" and python_version >= "3.7"
