### 2022-08-04 315346f
add notes on how to stop
