+++ Dockerfile
@@ -2,9 +2,12 @@
 # docker build --tag pymedphys .
 
 # Run this image with:
-# docker run --publish 8501:8501 pymedphys
+# docker run --publish 8501:8501 --name pymedphys pymedphys
 
 # View the result by opening browser at http://localhost:8501
+
+# Stop the container by running the following in a separate terminal:
+# docker stop pymedphys
 
 FROM python:3.10-slim
 
