### 2022-08-04 pymedphys/pymedphys#1717 @SimonBiggs
Document "Automated Tests"
> ... @SimonBiggs I see that most of the contributions in lib\pymedphys\tests came from you and other past contributors... I can try having a look at the tests, but I think it will require time and guidance.

_Originally posted by @nlambriICH in https://github.com/pymedphys/pymedphys/issues/1701#issuecomment-1195115931_

Referring to the review checklist:

>  Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

https://github.com/openjournals/joss-reviews/issues/4555#issuecomment-1187692538
