### 2022-08-04 pymedphys/pymedphys#1717 @SimonBiggs
Essentially, just need a document that says, to run the tests, run `poetry run pymedphys dev tests`
