### 2022-08-05 47cbdfd
re-introduce electron API by reverting commit 2466f6d1abc8eb0d4f6fa17e320373c95e261a94
