+++ lib/pymedphys/electronfactors.py
@@ -1,5 +1,5 @@
 """A suite of functions that model electron insert/cutout factors by
-parameterising them as equivalent ellipses.
+parametrising them as equivalent ellipses.
 """
 
 # pylint: disable = unused-import
