### 2022-08-05 5164638
Add type map of Mosaiq  Schedule table
