+++ lib/pymedphys/tests/mosaiq/data/types_map.toml
@@ -525,3 +525,49 @@ WobblingRadius_X = "decimal"
 WobblingRadius_Y = "decimal"
 LayercGrayPerMeterset = "decimal"
 RowVers = "timestamp"
+
+[Schedule]
+Sch_Id = "int"
+Sch_Set_Id = "int"
+Version = "smallint"
+Pat_ID1 = "int"
+App_DtTm = "datetime"
+Activity = "varchar"
+Location = "int"
+Staff_ID = "int"
+Inst_ID = "int"
+PCI_ID = "int"
+Alert = "smallint"
+Notes = "varchar"
+CHG_ID = "int"
+Wave_ID = "int"
+Private = "smallint"
+Create_DtTm = "datetime"
+Create_ID = "int"
+Edit_DtTm = "datetime"
+Edit_ID = "int"
+Sanct_DtTm = "datetime"
+Sanct_ID = "int"
+Sanct_DtTm2 = "datetime"
+Sanct_Id2 = "int"
+Note_ID = "int"
+Pau_ID = "int"
+Multi_Staff = "bit"
+NoCopay = "bit"
+Suppressed = "bit"
+Duration_time = "int"
+SCH_RelatedTo = "int"
+LowRange = "int"
+LowRangeType = "smallint"
+HighRange = "int"
+HighRangeType = "smallint"
+SEC_ID = "int"
+Series_Instance_ID = "int"
+SchStatus_Hist_UD = "varchar"
+SchStatus_Hist_SD = "varchar"
+ESI_Source = "varchar"
+ESI_AppointmentID = "varchar"
+ESI_LockMode = "tinyint"
+EQQ_ID = "int"
+OBR_ID = "int"
+RowVers = "datetime"
