+++ lib/pymedphys/tests/mosaiq/data/types_map.toml
@@ -570,4 +570,4 @@ ESI_AppointmentID = "varchar"
 ESI_LockMode = "tinyint"
 EQQ_ID = "int"
 OBR_ID = "int"
-RowVers = "datetime"
+RowVers = "timestamp"
