+++ lib/pymedphys/electronfactors.py
@@ -1,5 +1,5 @@
 """A suite of functions that model electron insert/cutout factors by
-parametrising them as equivalent ellipses.
+parameterising them as equivalent ellipses.
 """
 
 # pylint: disable = unused-import
@@ -9,7 +9,7 @@ from ._electronfactors import (
     calculate_percent_prediction_differences,
     convert2_ratio_perim_area,
     create_transformed_mesh,
-    parametrise_insert,
+    parameterise_insert,
     plot_model,
     spline_model,
     spline_model_with_deformability,
