### 2022-08-09 pymedphys/pymedphys#1697 @crcrewso
I think this is now closed with pr #1705
