### 2022-08-09 pymedphys/pymedphys#1717 @crcrewso
I think it's also time to include a requirements-dev even if it just references requirements.txt, poetry, and pandoc.
