### 2022-08-11 2247482
Add helper to get column data types
