### 2022-08-15 #4555 @ProfLeao
**About the text:**

In the Statement of need section, isolated _in-house development efforts_ are mentioned several times. I understand that initiatives of this nature are common in several Medical Physics services, but as it is a scientific publication, I believe that the statement should be endorsed by some bibliographic source or by examples that make it more consistent.
Do the authors believe this would be possible?
---------------
Sorry for the delay in continuing the review, I've been working away for several days.
