### 2022-08-18 pymedphys/pymedphys#1717 @nlambriICH
Hi guys, I took some days off, but I wanted to contribute to this issue as it was opened due to my comment.

I have added a section in the readme, as in the link posted by @SimonBiggs.

Let me know what you think in the PR I've opened. 😃
