### 2022-08-23 9d5fed9
Bump nbconvert from 6.5.0 to 6.5.1
Bumps [nbconvert](https://github.com/jupyter/nbconvert) from 6.5.0 to 6.5.1.
- [Release notes](https://github.com/jupyter/nbconvert/releases)
- [Commits](https://github.com/jupyter/nbconvert/compare/6.5...6.5.1)

---
updated-dependencies:
- dependency-name: nbconvert
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
