+++ requirements-all.txt
@@ -92,7 +92,7 @@ myst-nb==0.13.2; python_version >= "3.7"
 myst-parser==0.15.2; python_version >= "3.7"
 natsort==8.1.0; python_version >= "3.6"
 nbclient==0.5.13; python_full_version >= "3.7.0" and python_version >= "3.7"
-nbconvert==6.5.0; python_version >= "3.7"
+nbconvert==6.5.1; python_version >= "3.7"
 nbdime==3.1.1; python_version >= "3.7"
 nbformat==5.4.0; python_full_version >= "3.7.0" and python_version >= "3.7"
 nest-asyncio==1.5.5; python_full_version >= "3.7.0" and python_version >= "3.7"
