+++ requirements.txt
@@ -59,7 +59,7 @@ matplotlib==3.5.2; python_version >= "3.7"
 mistune==0.8.4; python_version >= "3.7"
 natsort==8.1.0; python_version >= "3.6"
 nbclient==0.5.13; python_full_version >= "3.7.0" and python_version >= "3.7"
-nbconvert==6.5.0; python_version >= "3.7"
+nbconvert==6.5.1; python_version >= "3.7"
 nbformat==5.4.0; python_full_version >= "3.7.0" and python_version >= "3.7"
 nest-asyncio==1.5.5; python_full_version >= "3.7.0" and python_version >= "3.7"
 networkx==2.8.5; python_version >= "3.8"
