+++ paper.bib
@@ -169,3 +169,143 @@
   url={https://iopscience.iop.org/article/10.1088/1361-6560/ac8269/meta},
   doi={10.1088/1361-6560/ac8269}
 }
+
+@article{kuo2020JACMP,
+  title={Implementation and validation of an in-house geometry optimization software for SRS VMAT planning of multiple cranial metastases},
+  author={Kuo, LiCheng and Zhang, PengPeng and Pham, Hai and Ballangrud, AM},
+  journal={Journal of Applied Clinical Medical Physics},
+  year={2020},
+  publisher={Wiley},
+  url={https://aapm.onlinelibrary.wiley.com/doi/full/10.1002/acm2.12961},
+  doi={10.1002/acm2.12961}
+}
+
+@article{Maughan2019MP,
+  title={Validation of post-treatment PET-based dosimetry software for hepatic radioembolization of Yttrium-90 microspheres},
+  author={Maughan, NM and Garcia-Ramirez, J and Arpidone, M and Swallen, A and Laforest, R and Goddu, SM and Parikh, PJ and Zoberi, JE},
+  journal={Medical Physics},
+  year={2019},
+  publisher={Wiley},
+  url={https://aapm.onlinelibrary.wiley.com/doi/full/10.1002/mp.13444},
+  doi={10.1002/mp.13444}
+}
+
+@article{Arumugam2016MP,
+  title={An online x-ray based position validation system for prostate hypofractionated radiotherapy},
+  author={Arumugam, S and Sidhom, M and Xing, A and Holloway, L},
+  journal={Medical Physics},
+  year={2016},
+  publisher={Wiley},
+  url={https://aapm.onlinelibrary.wiley.com/doi/full/10.1118/1.4940351},
+  doi={10.1118/1.4940351}
+}
+
+@article{Edvardsson2018PMB,
+  title={Motion induced interplay effects for VMAT radiotherapy},
+  author={Edvardsson, A and Nordström, F and Ceberg, C and Ceberg, S},
+  journal={Physics in Medicine and Biology},
+  year={2018},
+  publisher={IOP Publishing},
+  url={https://iopscience.iop.org/article/10.1088/1361-6560/aab957/meta},
+  doi={10.1088/1361-6560/aab957}
+}
+
+@article{LiJS2010MP,
+  title={Uncertainties in IMRT dosimetry},
+  author={Li, JS and Lin, T and Chen, L and Price, RA Jr and Ma, CM},
+  journal={Medical Physics},
+  year={2010},
+  publisher={Wiley},
+  url={https://aapm.onlinelibrary.wiley.com/doi/full/10.1118/1.3413997},
+  doi={10.1118/1.3413997}
+}
+
+@article{Bakhtiari2011MP,
+  title={Using an EPID for patient-specific VMAT quality assurance},
+  author={Bakhtiari, M and Kumaraswamy, L and Bailey, DW and de Boer, S and Malhotra, HK and Podgorsak, MB},
+  journal={Medical Physics},
+  year={2011},
+  publisher={Wiley},
+  url={https://aapm.onlinelibrary.wiley.com/doi/full/10.1118/1.3552925},
+  doi={10.1118/1.3552925}
+}
+
+@article{Latala2020MD,
+  title={The comparison of VMAT test results for Clinac 2300C/D and TrueBeam accelerators},
+  author={Latala, A and Fujak, E and Walewska, A and Kukołowicz, P},
+  journal={Medical Dosimetry},
+  year={2020},
+  publisher={Elsevier},
+  url={https://doi.org/10.1016/j.meddos.2019.12.007},
+  doi={10.1016/j.meddos.2019.12.007}
+}
+
+@article{Bhagroo2019MP,
[…]
