+++ paper.bib
@@ -276,7 +276,7 @@
   journal={Radiotherapy and Oncology},
   year={2019},
   publisher={Elsevier},
-  url={https://doi.org/10.1016/j.radonc.2019.07.007}
+  url={https://doi.org/10.1016/j.radonc.2019.07.007},
   doi={10.1016/j.radonc.2019.07.007}
 }
 
