### 2022-09-04 34db8f6
add test feet-in-supine in test_gamma_shell.py
