### 2022-09-06 da06169
Update lib/pymedphys/tests/gamma/test_gamma_shell.py
Co-authored-by: Matthew Jennings <Centrus.007@gmail.com>
