+++ lib/pymedphys/tests/gamma/test_gamma_shell.py
@@ -227,7 +227,7 @@ def test_if_descending_coordinate_in_DICOM():
     reference_filepath = pymedphys.data_path("RD.TBB_feet_in_supine.dcm")
     evaluation_filepath = pymedphys.data_path("RD.TBC_feet_in_supine.dcm")
 
-    reference = pydicom.read_file(str(reference_filepath), force=True)
+    reference = pydicom.dcmread(reference_filepath, force=True)
     evaluation = pydicom.read_file(str(evaluation_filepath), force=True)
 
     axes_reference, dose_reference = pymedphys.dicom.zyx_and_dose_from_dataset(
