### 2022-09-06 32a4cc7
Update lib/pymedphys/tests/gamma/test_gamma_shell.py
Co-authored-by: Matthew Jennings <Centrus.007@gmail.com>
