### 2022-09-10 #4555 @ProfLeao
Hello @osorensen  the `recommend-accept` command is only for editors right? I have finished my review process, what should I do?
