### 2022-09-11 #4555 @Matthew-Jennings
> @Matthew-Jennings, just to keep myself updated: do I understand correctly that you have attempted to address all the issues raised by the reviewers so far?

Almost! The outstanding items are:

1. An open PR with the statement of need on its own page but directly linked to in the opening paragraph of the README. This will likely be pulled in very soon. We hope this is acceptable to you, @gbaltz. If it's not too much trouble, we'd would prefer not to have the statement of need text directly in the README.
2. We haven't yet addressed @gbaltz's suggestion to include an examples section.

Hi, @gbaltz, just on your point regarding an examples section:

We've made the stylistic decision to maintain a relatively lean README and to treat it as a nearly pure dashboard from which the reader then navigates to the appropriate section/s of the documentation. One reason for this is that the PyMedPhys target audience is as much to potential contributors as it is to users, and some of the appropriate documentation for each is quite different. So, a section with examples of how to use PyMedPhys is perhaps not *as* relevant to potential contributors and vice versa.

The closest thing we have to examples for now is our "how-to guides" in the Users section. These are directly mentioned and linked to under the documentation section of the README. Does this suffice? Or do you have specific recommendations to streamline the reader's experience, given what I've mentioned above? Please note we are very open to such suggestions.
