### 2022-09-13 #4555 @gbaltz
> The closest thing we have to examples for now is our "how-to guides" in the Users section. These are directly mentioned and linked to under the documentation section of the README. Does this suffice? Or do you have specific recommendations to streamline the reader's experience, given what I've mentioned above? Please note we are very open to such suggestions.

@Matthew-Jennings yes this would be fine with me
