### 2022-09-15 #4555 @osorensen
@Matthew-Jennings, could you please make the following changes to the paper?

Lines 85, 105, and 135: change "monte carlo" to "Monte Carlo".
Line 97: change "winston-lutz" to "Winston-Lutz".
