+++ paper.bib
@@ -108,7 +108,7 @@
 }
 
 @article{gajewski2021commissioning,
-  title={Commissioning of GPU--accelerated monte carlo code FRED for clinical applications in proton therapy},
+  title={Commissioning of GPU--accelerated Monte Carlo code FRED for clinical applications in proton therapy},
   author={Gajewski, Jan and Garbacz, Magdalena and Chang, Chih-Wei and Czerska, Katarzyna and Durante, Marco and Krah, Nils and Krzempek, Katarzyna and Kope{\'c}, Renata and Lin, Liyong and Moj{\.z}eszek, Natalia and others},
   journal={Frontiers in Physics},
   volume={8},
