+++ paper.bib
@@ -108,7 +108,7 @@
 }
 
 @article{gajewski2021commissioning,
-  title={Commissioning of GPU--accelerated Monte Carlo code FRED for clinical applications in proton therapy},
+  title={Commissioning of GPU--accelerated {Monte} {Carlo} code FRED for clinical applications in proton therapy},
   author={Gajewski, Jan and Garbacz, Magdalena and Chang, Chih-Wei and Czerska, Katarzyna and Durante, Marco and Krah, Nils and Krzempek, Katarzyna and Kope{\'c}, Renata and Lin, Liyong and Moj{\.z}eszek, Natalia and others},
   journal={Frontiers in Physics},
   volume={8},
@@ -119,7 +119,7 @@
 }
 
 @article{lysakovski2021development,
-  title={Development and Benchmarking of a Monte Carlo Dose Engine for Proton Radiation Therapy},
+  title={Development and Benchmarking of a {Monte} {Carlo} Dose Engine for Proton Radiation Therapy},
   author={Lysakovski, Peter and Ferrari, Alfredo and Tessonnier, Thomas and Besuglow, Judith and Kopp, Benedikt and Mein, Stewart and Haberer, Thomas and Debus, J{\"u}rgen and Mairani, Andrea},
   journal={Frontiers in Physics},
   pages={655},
@@ -143,7 +143,7 @@
 }
 
 @article{douglass2021deepwl,
-  title={DeepWL: Robust EPID based Winston-Lutz analysis using deep learning, synthetic image generation and optical path-tracing},
+  title={DeepWL: Robust EPID based {Winston}-{Lutz} analysis using deep learning, synthetic image generation and optical path-tracing},
   author={Douglass, Michael John James and Keal, James Alan},
   journal={Physica Medica},
   volume={89},
@@ -241,7 +241,7 @@
 }
 
 @article{Bhagroo2019MP,
-  title={Secondary monitor unit calculations for VMAT using parallelized Monte Carlo simulations},
+  title={Secondary monitor unit calculations for VMAT using parallelized {Monte} {Carlo} simulations},
   author={Bhagroo, S and French, SB and Mathews, JA and Nazareth, DP},
   journal={Journal of Applied Clinical Medical Physics},
   year={2019},
