### 2022-09-15 pymedphys/pymedphys#1735 @SimonBiggs
Thanks @Matthew-Jennings for continuing to spearhead the paper submission 🙂

I had my ear drum burst today... And I suspect I'll continue to be out of action for the next few days.

@pchlap helped me with the most recent one, he might be in a position to help? Also, there was a recent poetry update to version 1.2.0. I think it might be worth updating to 1.2.0 locally, running a `poetry update` with the new version and updating the version number within the GitHub CI before making the release.

The biggest issue before the release will be getting to the bottom of the updates CI failure. That will need to be done first.
