### 2022-09-15 pymedphys/pymedphys#1735 @pchlap
Hi @SimonBiggs & @Matthew-Jennings,

Also agree, thanks @Matthew-Jennings and all for moving the paper submission forward!

I'm definitely happy to work on preparing the pymedphys release. I'll also update poetry first as @SimonBiggs suggests.

@SimonBiggs I don't suppose we had prepared any docs on the release process last time did we? I don't think we did so I'll go back and check what we did to start preparing the next release. I may need you (or someone else) to help remind me of a step or two along the way. I'll try to document as I go as well. :)

P.S. @SimonBiggs burst eardrum does not sound fun :( Hope you feel better soon!
