+++ paper.bib
@@ -55,7 +55,7 @@
 }
 
 @article{baltz2021validation,
-  title={Validation of spline modeling for calculation of electron insert factors for varian linear accelerators},
+  title={Validation of spline modeling for calculation of electron insert factors for {Varian} linear accelerators},
   author={Baltz, Garrett C and Kirsner, Steven M},
   journal={Journal of Applied Clinical Medical Physics},
   volume={22},
@@ -130,7 +130,7 @@
 
 @article{castle2022,
   author={Castle, James R and Duan, Jingwei and Feng, Xue and Chen, Quan},
-  title={Development of a virtual source model for Monte Carlo-based independent dose calculation for varian linac},
+  title={Development of a virtual source model for {Monte} {Carlo}-based independent dose calculation for varian linac},
   publisher={Wiley Online Library},
   journal={Journal of Applied Clinical Medical Physics},
   pages={e13556},
@@ -181,7 +181,7 @@
 }
 
 @article{Maughan2019MP,
-  title={Validation of post-treatment PET-based dosimetry software for hepatic radioembolization of Yttrium-90 microspheres},
+  title={Validation of post-treatment PET-based dosimetry software for hepatic radioembolization of {Yttrium}-90 microspheres},
   author={Maughan, NM and Garcia-Ramirez, J and Arpidone, M and Swallen, A and Laforest, R and Goddu, SM and Parikh, PJ and Zoberi, JE},
   journal={Medical Physics},
   year={2019},
@@ -231,7 +231,7 @@
 }
 
 @article{Latala2020MD,
-  title={The comparison of VMAT test results for Clinac 2300C/D and TrueBeam accelerators},
+  title={The comparison of VMAT test results for Clinac {2300C/D} and TrueBeam accelerators},
   author={Latala, A and Fujak, E and Walewska, A and Kukołowicz, P},
   journal={Medical Dosimetry},
   year={2020},
