### 2022-09-18 bad14d7
avoid crashing when --pseudo is not provided via experimental fallback to using release anonymisation strategy and keywords
