### 2022-09-18 9aa4b48
added test to run cli (directory input/output) pymedphys --verbose experimental dicom anonymise without the --pseudo
