### 2022-09-19 pymedphys/pymedphys#1735 @Matthew-Jennings
Thanks, @pchlap, that would be great!

If you do feel like documenting along the way, I think a useful place might be a new docs page (e.g. "Making a release") under _Contributors Guide / Repository Information_. Even a very simply formatted dot point list would be quite helpful, I think.
