+++ .github/workflows/mosaiq-db-tests.yml
@@ -65,7 +65,7 @@ jobs:
       - name: Install Poetry
         uses: snok/install-poetry@v1.1.1
         with:
-          version: 1.1.13
+          version: 1.2.1
           virtualenvs-create: true
 
       - name: Configure Poetry to not use experimental installer
