+++ .github/workflows/library.yml
@@ -173,7 +173,7 @@ jobs:
       - name: Install Poetry
         uses: snok/install-poetry@v1
         with:
-          version: 1.1.13
+          version: 1.2.1
           virtualenvs-create: true
 
       - name: Update PATH in Windows
