### 2022-09-19 d31be47
Change snok/install-poetry to v1 for mosaiq tests
