+++ .github/workflows/mosaiq-db-tests.yml
@@ -63,7 +63,7 @@ jobs:
           pip install --upgrade pip setuptools
 
       - name: Install Poetry
-        uses: snok/install-poetry@v1.1.1
+        uses: snok/install-poetry@v1
         with:
           version: 1.2.1
           virtualenvs-create: true
