### 2022-09-20 4774dfd
Explicitly install tests within doctests
