### 2022-09-20 e09d242
Install tests in both tests and doctests
