### 2022-09-20 f476350
Upper pin pandas and streamlit
