+++ requirements-cli.txt
@@ -1,2 +1,2 @@
-toml==0.10.2; (python_version >= "2.6" and python_full_version < "3.0.0") or (python_full_version >= "3.3.0")
-typing-extensions==4.3.0; python_version >= "3.7"
+toml==0.10.2 ; python_version >= "3.8" and python_version < "3.11"
+typing-extensions==4.3.0 ; python_version >= "3.8" and python_version < "3.11"
