+++ requirements-all.txt
@@ -1,230 +1,234 @@
-alabaster==0.7.12; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
-altair==4.2.0; python_version >= "3.7"
-anyio==3.6.1; python_full_version >= "3.6.2" and python_version >= "3.7"
-appnope==0.1.3; sys_platform == "darwin" and python_version >= "3.8" and platform_system == "Darwin"
-argon2-cffi-bindings==21.2.0; python_version >= "3.7"
-argon2-cffi==21.3.0; python_version >= "3.7"
-argue==0.3.1
-astroid==2.11.7; python_full_version >= "3.6.2"
-asttokens==2.0.5; python_version >= "3.8"
-atomicwrites==1.4.1; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.4.0"
-attrs==21.4.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
-babel==2.10.3; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
-backcall==0.2.0; python_version >= "3.8"
-backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.7" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
-beautifulsoup4==4.11.1; python_full_version >= "3.6.0" and python_version >= "3.7"
-black==22.6.0; python_full_version >= "3.6.2"
-bleach==5.0.1; python_version >= "3.7"
-blinker==1.5; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
-cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0"
-certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
-cffi==1.15.1; sys_platform == "linux" and python_version >= "3.7" and implementation_name == "pypy"
-cfgv==3.3.1; python_full_version >= "3.6.1" and python_version >= "3.7"
-charset-normalizer==2.1.0; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
-click==8.1.3; python_version >= "3.7" and python_full_version >= "3.6.2" and python_version < "4.0"
-colorama==0.4.5; python_full_version >= "3.7.2" and platform_system == "Windows" and sys_platform == "win32" and python_version >= "3.8" and (python_version >= "2.7" and python_full_version < "3.0.0" and platform_system == "Windows" or python_full_version >= "3.5.0" and platform_system == "Windows") and (python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" or python_full_version >= "3.5.0" and python_version >= "3.7" and sys_platform == "win32") and (python_version >= "3.8" and python_full_version < "3.0.0" and sys_platform == "win32" or sys_platform == "win32" and python_version >= "3.8" and python_full_version >= "3.5.0") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7")
[…]
