+++ requirements.txt
@@ -1,150 +1,106 @@
-altair==4.2.0; python_version >= "3.7"
-appnope==0.1.3; sys_platform == "darwin" and python_version >= "3.8" and platform_system == "Darwin"
-argon2-cffi-bindings==21.2.0; python_version >= "3.7"
-argon2-cffi==21.3.0; python_version >= "3.7"
-argue==0.3.1
-asttokens==2.0.5; python_version >= "3.8"
-attrs==21.4.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
-backcall==0.2.0; python_version >= "3.8"
-backports.zoneinfo==0.2.1; python_version < "3.9" and python_version >= "3.7" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
-beautifulsoup4==4.11.1; python_full_version >= "3.6.0" and python_version >= "3.7"
-bleach==5.0.1; python_version >= "3.7"
-blinker==1.5; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
-cachetools==5.2.0; python_version >= "3.7" and python_version < "4.0"
-certifi==2022.6.15; python_version >= "3.7" and python_version < "4"
-cffi==1.15.1; sys_platform == "linux" and python_version >= "3.7" and implementation_name == "pypy"
-charset-normalizer==2.1.0; python_version >= "3.7" and python_version < "4" and python_full_version >= "3.6.0"
-click==8.1.3; python_version >= "3.7"
-colorama==0.4.5; python_version >= "3.8" and python_full_version < "3.0.0" and platform_system == "Windows" and sys_platform == "win32" or python_full_version >= "3.5.0" and platform_system == "Windows" and sys_platform == "win32" and python_version >= "3.8"
-commonmark==0.9.1; python_full_version >= "3.6.3" and python_full_version < "4.0.0" and python_version >= "3.7"
-cryptography==37.0.4; sys_platform == "linux" and python_version >= "3.7"
-cssselect==1.1.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
-cssutils==2.5.1; python_version >= "3.7"
-cycler==0.11.0; python_version >= "3.7"
-dbfread==2.0.7
-debugpy==1.6.2; python_version >= "3.7"
-decorator==5.1.1; python_version >= "3.8"
-defusedxml==0.7.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.7"
-dicompyler-core==0.5.5
-entrypoints==0.4; python_version >= "3.7"
-executing==0.9.1; python_version >= "3.8"
-fastjsonschema==2.16.1; python_version >= "3.7"
-fonttools==4.34.4; python_version >= "3.7"
-fsspec==2022.7.1; python_version >= "3.7"
-gitdb==4.0.9; python_version >= "3.7"
-gitpython==3.1.27; python_version >= "3.7"
[…]
