### 2022-09-20 52f4d3a
Explicit boolean operation for doctests
