### 2022-09-21 f803b68
Un-upper pin streamlit, pandas and upper pin pylint
