+++ pyproject.toml
@@ -71,14 +71,14 @@ dicompyler-core = { version = "*", optional = true }   # groups = ["user", "all"
 scikit-learn = { version = "*", optional = true }      # groups = ["user", "all", "mosaiq"]
 imageio = { version = "*", optional = true }           # groups = ["user", "all"]
 
-streamlit = { version = "<1.12.0", optional = true }     # groups = ["user", "all"]
+streamlit = { version = "*", optional = true }     # groups = ["user", "all"]
 streamlit-ace = { version = "*", optional = true } # groups = ["user", "all"]
 
 # NumPy's lower bound here is for typing support
 numpy = { version = ">=1.20.2", optional = true } # groups = ["user", "all", "icom", "docs"]
 
 # The following lower bounded packages are due to certain APIs utilised
-pandas = { version = ">=1.0.0,<1.5.0", optional = true }  # groups = ["user", "all", "mosaiq"]
+pandas = { version = ">=1.0.0", optional = true }  # groups = ["user", "all", "mosaiq"]
 pydicom = { version = ">=2.0.0", optional = true } # groups = ["user", "all", "dicom", "docs"]
 
 
@@ -141,7 +141,7 @@ pytest-sugar = { version = "*", optional = true }         # groups = ["all", "te
 hypothesis = { version = "<6", optional = true }          # groups = ["all", "tests"]
 astroid = { version = "*", optional = true }              # groups = ["all", "tests"]
 psutil = { version = "*", optional = true }               # groups = ["all", "tests"]
-pylint = { version = "*", optional = true }               # groups = ["all", "tests"]
+pylint = { version = "<2.15.0", optional = true }               # groups = ["all", "tests"]
 pytest-rerunfailures = { version = "*", optional = true } # groups = ["all", "tests"]
 
 # ---------------- #
