+++ requirements-deploy.txt
@@ -1,6 +1,6 @@
 altair==4.2.0 ; python_version >= "3.8" and python_version < "3.11"
 argue==0.3.1 ; python_version >= "3.8" and python_version < "3.11"
-astroid==2.12.10 ; python_version >= "3.8" and python_version < "3.11"
+astroid==2.11.7 ; python_version >= "3.8" and python_version < "3.11"
 attrs==21.4.0 ; python_version >= "3.8" and python_version < "3.11"
 backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
 blinker==1.5 ; python_version >= "3.8" and python_version < "3.11"
@@ -21,7 +21,7 @@ decorator==5.1.1 ; python_version >= "3.8" and python_version < "3.11"
 dicompyler-core==0.5.5 ; python_version >= "3.8" and python_version < "3.11"
 dill==0.3.5.1 ; python_version >= "3.8" and python_version < "3.11"
 entrypoints==0.4 ; python_version >= "3.8" and python_version < "3.11"
-fonttools==4.37.2 ; python_version >= "3.8" and python_version < "3.11"
+fonttools==4.37.3 ; python_version >= "3.8" and python_version < "3.11"
 fsspec==2022.8.2 ; python_version >= "3.8" and python_version < "3.11"
 gitdb==4.0.9 ; python_version >= "3.8" and python_version < "3.11"
 gitpython==3.1.27 ; python_version >= "3.8" and python_version < "3.11"
@@ -47,9 +47,9 @@ mccabe==0.7.0 ; python_version >= "3.8" and python_version < "3.11"
 more-itertools==8.14.0 ; python_version >= "3.8" and python_version < "3.11"
 natsort==8.2.0 ; python_version >= "3.8" and python_version < "3.11"
 networkx==2.8.6 ; python_version >= "3.8" and python_version < "3.11"
-numpy==1.23.3 ; python_version < "3.11" and python_version >= "3.8"
+numpy==1.23.3 ; python_version >= "3.8" and python_version < "3.11"
 packaging==21.3 ; python_version >= "3.8" and python_version < "3.11"
-pandas==1.4.4 ; python_version >= "3.8" and python_version < "3.11"
+pandas==1.5.0 ; python_version >= "3.8" and python_version < "3.11"
 pillow==9.2.0 ; python_version >= "3.8" and python_version < "3.11"
 platformdirs==2.5.2 ; python_version >= "3.8" and python_version < "3.11"
 plotly==5.10.0 ; python_version >= "3.8" and python_version < "3.11"
@@ -64,7 +64,7 @@ pydeck==0.8.0b3 ; python_version >= "3.8" and python_version < "3.11"
 pydicom==2.3.0 ; python_version >= "3.8" and python_version < "3.11"
 pygments==2.13.0 ; python_version >= "3.8" and python_version < "3.11"
 pylinac==2.3.2 ; python_version >= "3.8" and python_version < "3.11"
-pylint==2.15.3 ; python_version >= "3.8" and python_version < "3.11"
+pylint==2.14.5 ; python_version >= "3.8" and python_version < "3.11"
 pymedphys-pylibjpeg-libjpeg==1.3.1 ; python_version >= "3.8" and python_version < "3.11"
 pympler==1.0.1 ; python_version >= "3.8" and python_version < "3.11"
 pymssql==2.2.5 ; python_version >= "3.8" and python_version < "3.11"
@@ -97,7 +97,7 @@ smmap==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
 sortedcontainers==2.4.0 ; python_version >= "3.8" and python_version < "3.11"
[…]
