+++ requirements-docs.txt
@@ -23,7 +23,7 @@ docutils==0.17.1 ; python_version >= "3.8" and python_version < "3.11"
 entrypoints==0.4 ; python_version >= "3.8" and python_version < "3.11"
 executing==1.0.0 ; python_version >= "3.8" and python_version < "3.11"
 fastjsonschema==2.16.2 ; python_version >= "3.8" and python_version < "3.11"
-fonttools==4.37.2 ; python_version >= "3.8" and python_version < "3.11"
+fonttools==4.37.3 ; python_version >= "3.8" and python_version < "3.11"
 gitdb==4.0.9 ; python_version >= "3.8" and python_version < "3.11"
 gitpython==3.1.27 ; python_version >= "3.8" and python_version < "3.11"
 greenlet==1.1.3 ; python_version >= "3.8" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "3.11"
@@ -109,7 +109,7 @@ sphinx-comments==0.0.3 ; python_version >= "3.8" and python_version < "3.11"
 sphinx-copybutton==0.5.0 ; python_version >= "3.8" and python_version < "3.11"
 sphinx-design==0.1.0 ; python_version >= "3.8" and python_version < "3.11"
 sphinx-external-toc==0.2.4 ; python_version >= "3.8" and python_version < "3.11"
-sphinx-jupyterbook-latex==0.4.6 ; python_version >= "3.8" and python_version < "3.11"
+sphinx-jupyterbook-latex==0.4.7 ; python_version >= "3.8" and python_version < "3.11"
 sphinx-multitoc-numbering==0.1.3 ; python_version >= "3.8" and python_version < "3.11"
 sphinx-thebe==0.1.2 ; python_version >= "3.8" and python_version < "3.11"
 sphinx-togglebutton==0.3.2 ; python_version >= "3.8" and python_version < "3.11"
