+++ requirements-all.txt
@@ -5,7 +5,7 @@ appnope==0.1.3 ; python_version >= "3.8" and python_version < "3.11" and sys_pla
 argon2-cffi-bindings==21.2.0 ; python_version >= "3.8" and python_version < "3.11"
 argon2-cffi==21.3.0 ; python_version >= "3.8" and python_version < "3.11"
 argue==0.3.1 ; python_version >= "3.8" and python_version < "3.11"
-astroid==2.12.10 ; python_version >= "3.8" and python_version < "3.11"
+astroid==2.11.7 ; python_version >= "3.8" and python_version < "3.11"
 asttokens==2.0.8 ; python_version >= "3.8" and python_version < "3.11"
 attrs==21.4.0 ; python_version >= "3.8" and python_version < "3.11"
 babel==2.10.3 ; python_version >= "3.8" and python_version < "3.11"
@@ -41,7 +41,7 @@ entrypoints==0.4 ; python_version >= "3.8" and python_version < "3.11"
 executing==1.0.0 ; python_version >= "3.8" and python_version < "3.11"
 fastjsonschema==2.16.2 ; python_version >= "3.8" and python_version < "3.11"
 filelock==3.8.0 ; python_version >= "3.8" and python_version < "3.11"
-fonttools==4.37.2 ; python_version >= "3.8" and python_version < "3.11"
+fonttools==4.37.3 ; python_version >= "3.8" and python_version < "3.11"
 fsspec==2022.8.2 ; python_version >= "3.8" and python_version < "3.11"
 gitdb==4.0.9 ; python_version >= "3.8" and python_version < "3.11"
 gitpython==3.1.27 ; python_version >= "3.8" and python_version < "3.11"
@@ -101,9 +101,9 @@ nest-asyncio==1.5.5 ; python_version >= "3.8" and python_version < "3.11"
 networkx==2.8.6 ; python_version >= "3.8" and python_version < "3.11"
 nodeenv==1.7.0 ; python_version >= "3.8" and python_version < "3.11"
 notebook==6.4.12 ; python_version >= "3.8" and python_version < "3.11"
-numpy==1.23.3 ; python_version < "3.11" and python_version >= "3.8"
+numpy==1.23.3 ; python_version >= "3.8" and python_version < "3.11"
 packaging==21.3 ; python_version >= "3.8" and python_version < "3.11"
-pandas==1.4.4 ; python_version >= "3.8" and python_version < "3.11"
+pandas==1.5.0 ; python_version >= "3.8" and python_version < "3.11"
 pandocfilters==1.5.0 ; python_version >= "3.8" and python_version < "3.11"
 parso==0.8.3 ; python_version >= "3.8" and python_version < "3.11"
 pathspec==0.10.1 ; python_version >= "3.8" and python_version < "3.11"
@@ -132,7 +132,7 @@ pydeck==0.8.0b3 ; python_version >= "3.8" and python_version < "3.11"
 pydicom==2.3.0 ; python_version >= "3.8" and python_version < "3.11"
 pygments==2.13.0 ; python_version >= "3.8" and python_version < "3.11"
 pylinac==2.3.2 ; python_version >= "3.8" and python_version < "3.11"
-pylint==2.15.3 ; python_version >= "3.8" and python_version < "3.11"
+pylint==2.14.5 ; python_version >= "3.8" and python_version < "3.11"
 pymedphys-pylibjpeg-libjpeg==1.3.1 ; python_version >= "3.8" and python_version < "3.11"
 pympler==1.0.1 ; python_version >= "3.8" and python_version < "3.11"
 pymssql==2.2.5 ; python_version >= "3.8" and python_version < "3.11"
[…]

[…]
