### 2022-09-26 pymedphys/pymedphys#1735 @Matthew-Jennings
> I think it might be worth updating to 1.2.0 locally, running a poetry update with the new version and updating the version number within the GitHub CI before making the release.

@SimonBiggs, given the issues faced in #1739, is there an alternative "patch" release pathway that we can take to avoid further delay in paper acceptance?
