### 2022-09-27 pymedphys/pymedphys#1735 @pchlap
> As long as poetry upgrade has been run without issue then we can leave poetry at v1.1.

I'd just add that I think the main problem I've been facing there is the upgrades of streamlit and pandas which are causing problems with pylint. I'd expect these to happen even without the upgrade to poetry.

At this stage the only easy solution to that is to upper pin them which isn't ideal. There do seem to be breaking changes for streamlit (pandas should be fine to just add in pylint disables). But I'm not very familiar with streamlit so I might be way off on this.
