### 2022-09-27 pymedphys/pymedphys#1735 @SimonBiggs
> At this stage the only easy solution to that is to upper pin them which isn't ideal.

Given streamlit is a part of our "application" and less a "library package"... I'd be okay for more leniency on being able to upper pin streamlit. If that gives you an easy fix, happy for you to run with that option for this release.
