+++ requirements.txt
@@ -4,7 +4,7 @@ attrs==21.4.0 ; python_version >= "3.8" and python_version < "3.11"
 backports-zoneinfo==0.2.1 ; python_version >= "3.8" and python_version < "3.9"
 blinker==1.5 ; python_version >= "3.8" and python_version < "3.11"
 cachetools==5.2.0 ; python_version >= "3.8" and python_version < "3.11"
-certifi==2022.9.14 ; python_version >= "3.8" and python_version < "3.11"
+certifi==2022.9.24 ; python_version >= "3.8" and python_version < "3.11"
 cffi==1.15.1 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "linux"
 charset-normalizer==2.1.1 ; python_version >= "3.8" and python_version < "3.11"
 click==8.1.3 ; python_version >= "3.8" and python_version < "3.11"
@@ -27,7 +27,7 @@ greenlet==1.1.3 ; python_version >= "3.8" and (platform_machine == "aarch64" or 
 idna==3.4 ; python_version >= "3.8" and python_version < "3.11"
 imageio==2.22.0 ; python_version >= "3.8" and python_version < "3.11"
 importlib-metadata==4.12.0 ; python_version >= "3.8" and python_version < "3.11"
-jaraco-classes==3.2.2 ; python_version >= "3.8" and python_version < "3.11"
+jaraco-classes==3.2.3 ; python_version >= "3.8" and python_version < "3.11"
 jeepney==0.8.0 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "linux"
 jinja2==3.1.2 ; python_version >= "3.8" and python_version < "3.11"
 joblib==1.2.0 ; python_version >= "3.8" and python_version < "3.11"
@@ -75,25 +75,25 @@ scipy==1.9.1 ; python_version >= "3.8" and python_version < "3.11"
 secretstorage==3.3.3 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "linux"
 semver==2.13.0 ; python_version >= "3.8" and python_version < "3.11"
 setuptools-scm==7.0.5 ; python_version >= "3.8" and python_version < "3.11"
-setuptools==65.3.0 ; python_version >= "3.8" and python_version < "3.11"
+setuptools==65.4.0 ; python_version >= "3.8" and python_version < "3.11"
 shapely==1.8.4 ; python_version >= "3.8" and python_version < "3.11"
 six==1.16.0 ; python_version >= "3.8" and python_version < "3.11"
 smmap==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
 sqlalchemy==1.4.41 ; python_version >= "3.8" and python_version < "3.11"
 streamlit-ace==0.1.1 ; python_version >= "3.8" and python_version < "3.11"
 streamlit==1.12.0 ; python_version >= "3.8" and python_version < "3.11"
-tenacity==8.0.1 ; python_version >= "3.8" and python_version < "3.11"
+tenacity==8.1.0 ; python_version >= "3.8" and python_version < "3.11"
 threadpoolctl==3.1.0 ; python_version >= "3.8" and python_version < "3.11"
 tifffile==2022.8.12 ; python_version >= "3.8" and python_version < "3.11"
 timeago==1.0.16 ; python_version >= "3.8" and python_version < "3.11"
 toml==0.10.2 ; python_version >= "3.8" and python_version < "3.11"
 tomli==2.0.1 ; python_version >= "3.8" and python_version < "3.11"
-tomlkit==0.11.4 ; python_version >= "3.8" and python_version < "3.11"
+tomlkit==0.11.5 ; python_version >= "3.8" and python_version < "3.11"
[…]
