+++ requirements-all.txt
@@ -16,7 +16,7 @@ black==22.8.0 ; python_version >= "3.8" and python_version < "3.11"
 bleach==5.0.1 ; python_version >= "3.8" and python_version < "3.11"
 blinker==1.5 ; python_version >= "3.8" and python_version < "3.11"
 cachetools==5.2.0 ; python_version >= "3.8" and python_version < "3.11"
-certifi==2022.9.14 ; python_version >= "3.8" and python_version < "3.11"
+certifi==2022.9.24 ; python_version >= "3.8" and python_version < "3.11"
 cffi==1.15.1 ; python_version >= "3.8" and python_version < "3.11"
 cfgv==3.3.1 ; python_version >= "3.8" and python_version < "3.11"
 charset-normalizer==2.1.1 ; python_version >= "3.8" and python_version < "3.11"
@@ -38,7 +38,7 @@ distlib==0.3.6 ; python_version >= "3.8" and python_version < "3.11"
 doc8==0.11.2 ; python_version >= "3.8" and python_version < "3.11"
 docutils==0.17.1 ; python_version >= "3.8" and python_version < "3.11"
 entrypoints==0.4 ; python_version >= "3.8" and python_version < "3.11"
-executing==1.0.0 ; python_version >= "3.8" and python_version < "3.11"
+executing==1.1.0 ; python_version >= "3.8" and python_version < "3.11"
 fastjsonschema==2.16.2 ; python_version >= "3.8" and python_version < "3.11"
 filelock==3.8.0 ; python_version >= "3.8" and python_version < "3.11"
 fonttools==4.37.3 ; python_version >= "3.8" and python_version < "3.11"
@@ -54,12 +54,12 @@ imagesize==1.4.1 ; python_version >= "3.8" and python_version < "3.11"
 importlib-metadata==4.12.0 ; python_version >= "3.8" and python_version < "3.11"
 importlib-resources==5.9.0 ; python_version >= "3.8" and python_version < "3.9"
 iniconfig==1.1.1 ; python_version >= "3.8" and python_version < "3.11"
-ipykernel==6.15.3 ; python_version >= "3.8" and python_version < "3.11"
+ipykernel==6.16.0 ; python_version >= "3.8" and python_version < "3.11"
 ipython-genutils==0.2.0 ; python_version >= "3.8" and python_version < "3.11"
 ipython==8.5.0 ; python_version >= "3.8" and python_version < "3.11"
 ipywidgets==7.7.2 ; python_version >= "3.8" and python_version < "3.11"
 isort==5.10.1 ; python_version >= "3.8" and python_version < "3.11"
-jaraco-classes==3.2.2 ; python_version >= "3.8" and python_version < "3.11"
+jaraco-classes==3.2.3 ; python_version >= "3.8" and python_version < "3.11"
 jedi==0.18.1 ; python_version >= "3.8" and python_version < "3.11"
 jeepney==0.8.0 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "linux"
 jinja2==3.1.2 ; python_version >= "3.8" and python_version < "3.11"
@@ -70,7 +70,7 @@ jupyter-cache==0.4.3 ; python_version >= "3.8" and python_version < "3.11"
 jupyter-client==7.3.5 ; python_version >= "3.8" and python_version < "3.11"
 jupyter-core==4.11.1 ; python_version >= "3.8" and python_version < "3.11"
 jupyter-server-mathjax==0.2.6 ; python_version >= "3.8" and python_version < "3.11"
-jupyter-server==1.18.1 ; python_version >= "3.8" and python_version < "3.11"
[…]
