+++ pyproject.toml
@@ -71,7 +71,7 @@ dicompyler-core = { version = "*", optional = true }   # groups = ["user", "all"
 scikit-learn = { version = "*", optional = true }      # groups = ["user", "all", "mosaiq"]
 imageio = { version = "*", optional = true }           # groups = ["user", "all"]
 
-streamlit = { version = "*", optional = true }     # groups = ["user", "all"]
+streamlit = { version = "<1.13.0", optional = true }     # groups = ["user", "all"]
 streamlit-ace = { version = "*", optional = true } # groups = ["user", "all"]
 
 # NumPy's lower bound here is for typing support

[…]
