+++ requirements.txt
@@ -81,7 +81,7 @@ six==1.16.0 ; python_version >= "3.8" and python_version < "3.11"
 smmap==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
 sqlalchemy==1.4.41 ; python_version >= "3.8" and python_version < "3.11"
 streamlit-ace==0.1.1 ; python_version >= "3.8" and python_version < "3.11"
-streamlit==1.12.0 ; python_version >= "3.8" and python_version < "3.11"
+streamlit==1.11.1 ; python_version >= "3.8" and python_version < "3.11"
 tenacity==8.1.0 ; python_version >= "3.8" and python_version < "3.11"
 threadpoolctl==3.1.0 ; python_version >= "3.8" and python_version < "3.11"
 tifffile==2022.8.12 ; python_version >= "3.8" and python_version < "3.11"
