+++ requirements-deploy.txt
@@ -97,7 +97,7 @@ smmap==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
 sortedcontainers==2.4.0 ; python_version >= "3.8" and python_version < "3.11"
 sqlalchemy==1.4.41 ; python_version >= "3.8" and python_version < "3.11"
 streamlit-ace==0.1.1 ; python_version >= "3.8" and python_version < "3.11"
-streamlit==1.12.0 ; python_version >= "3.8" and python_version < "3.11"
+streamlit==1.11.1 ; python_version >= "3.8" and python_version < "3.11"
 tenacity==8.1.0 ; python_version >= "3.8" and python_version < "3.11"
 termcolor==2.0.1 ; python_version >= "3.8" and python_version < "3.11"
 threadpoolctl==3.1.0 ; python_version >= "3.8" and python_version < "3.11"
