+++ requirements-all.txt
@@ -196,7 +196,7 @@ sqlalchemy==1.4.41 ; python_version >= "3.8" and python_version < "3.11"
 stack-data==0.5.1 ; python_version >= "3.8" and python_version < "3.11"
 stevedore==4.0.0 ; python_version >= "3.8" and python_version < "3.11"
 streamlit-ace==0.1.1 ; python_version >= "3.8" and python_version < "3.11"
-streamlit==1.12.0 ; python_version >= "3.8" and python_version < "3.11"
+streamlit==1.11.1 ; python_version >= "3.8" and python_version < "3.11"
 tabulate==0.8.10 ; python_version >= "3.8" and python_version < "3.11"
 tenacity==8.1.0 ; python_version >= "3.8" and python_version < "3.11"
 termcolor==2.0.1 ; python_version >= "3.8" and python_version < "3.11"
