### 2022-09-28 a61f8b3
Combine tests and doctests in CI
