+++ .github/workflows/library.yml
@@ -283,18 +283,12 @@ jobs:
           poetry install -E docs
 
       ## Install with Tests Environment
+      # TODO: Redesign CI to be more efficient once this issue lands python-poetry/poetry#4842
 
-      - name: Install a tests environment
+      - name: Install a tests/doctests environment
         if: matrix.task == 'tests' || matrix.task == 'cypress' || matrix.task == 'slow' || matrix.task == 'pyright'
         run: |
-          poetry install -E user -E tests
-
-      ## Install with Doctests Environment
-
-      - name: Install a doctests environment
-        if: matrix.os == 'ubuntu' && (matrix.task == 'tests' || matrix.task == 'pyright')
-        run: |
-          poetry install -E user -E doctests -E tests
+          poetry install -E user -E tests -E doctests
 
       ## Install remaining packages for pyright
 
