+++ pyproject.toml
@@ -87,11 +87,9 @@ pydicom = { version = ">=2.0.0", optional = true } # groups = ["user", "all", "d
 shapely = { version = ">=1.7.0", optional = true }       # groups = ["user", "all"]
 scikit-image = { version = ">=0.18.1", optional = true } # groups = ["user", "all"]
 
-# Utilising the pymedphys fork of pylibjpeg-libjpeg to fix the following issue:
+# Lower bound here is due to the following issue:
 # https://github.com/pymedphys/pymedphys/issues/1556#issuecomment-1152821285
-pymedphys-pylibjpeg-libjpeg = { version = "1.3.1", optional = true } # groups = ["user", "all"]
-# pylibjpeg-libjpeg = { version = "*", optional = true } # groups = ["user", "all"]
-
+pylibjpeg-libjpeg = { version = ">1.3", optional = true } # groups = ["user", "all"]
 
 # PyLinac is pinned due to historical algorithm changes with both minor and
 # patch version bumps.
