### 2022-09-29 5ab13c2
Update .github/workflows/mosaiq-db-tests.yml
Co-authored-by: Simon Biggs <simon.biggs@radiotherapy.ai>
