+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pymedphys"
-version = "0.38.0"
+version = "0.39.0"
 readme = "README.rst"
 description = "Medical Physics library"
 authors = [
