+++ lib/pymedphys/_version.py
@@ -1,5 +1,5 @@
 # DO NOT EDIT THIS FILE!
 # This file has been autogenerated by `poetry run pymedphys dev propagate`
 
-version_info = [0, 38, 0]
-__version__ = "0.38.0"
+version_info = [0, 39, 0]
+__version__ = "0.39.0"
