### 2022-09-29 #4555 @Matthew-Jennings
> * Make a tagged release of your software, and list the version tag of the archived version here.

Just to clarify: the new tagged released would be version 0.**39**.0. Is this the version that we should list here, archive in Zenodo and for which we should list the DOI? Or the previous version, 0.**38**.0)?

I'd note that a forthcoming 0.39.0 probably more closely matches what was reviewed, since virtually all of the changes in `main` made since the 0.38.0 release were made prior to commencement of the JOSS review, and we've made some small changes to address reviewer comments that will only be present in 0.39.0.
