+++ CHANGELOG.md
@@ -9,14 +9,22 @@ This project adheres to
 
 ## [0.39.0]
 
-### New features and enhancements
+### News around this release
 
 - This is the version of PyMedPhys reviewed for first publication in the
   [Journal of Open Source Software (JOSS)](https://joss.theoj.org/).
+
+### New features and enhancements
+
 - The Gamma tool now utilises the `logging` module. The `quiet` parameter for
   `pymedphys.gamma` is deprecated but its use is non-breaking (for now).
 - The `pymedphys.electronfactors` has been added back in to the public API in
   [Pull Request](https://github.com/pymedphys/pymedphys/pull/1719).
+
+### Non-API changing enhancements
+
+- The `dev` extra has been reduced and an additional extra `all` has been added:
+  [Pull Request](https://github.com/pymedphys/pymedphys/pull/1710)
 
 ## [0.38.0]
 
