+++ _sources/release-notes.md
@@ -7,12 +7,24 @@ All notable changes to are documented here.
 This project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.39.0]
+
+### News around this release
+
+- This is the version of PyMedPhys reviewed for first publication in the
+  [Journal of Open Source Software (JOSS)](https://joss.theoj.org/).
 
 ### New features and enhancements
 
 - The Gamma tool now utilises the `logging` module. The `quiet` parameter for
   `pymedphys.gamma` is deprecated but its use is non-breaking (for now).
+- The `pymedphys.electronfactors` has been added back in to the public API in
+  [Pull Request](https://github.com/pymedphys/pymedphys/pull/1719).
+
+### Non-API changing enhancements
+
+- The `dev` extra has been reduced and an additional extra `all` has been added:
+  [Pull Request](https://github.com/pymedphys/pymedphys/pull/1710)
 
 ## [0.38.0]
 
