### 2022-09-29 pymedphys/pymedphys#1735 @nlambriICH
Hi all and thanks for your efforts!

I just wanted to ask do we plan to add the authors' ORCID before final revision?
