### 2022-09-29 pymedphys/pymedphys#1735 @Matthew-Jennings
> Hi all and thanks for your efforts!
>
> I just wanted to ask do we plan to add the authors' ORCID before final revision?

Yes, good point.

@SimonBiggs, @sjswerdloff, @pchlap, @dg1an3, @rembishj, @JacobMcAloney, @kingrpaul, @ayalalazaro, @guanfada, @nlambriICH, @crcrewso, @msobolewski: if any of you would like to add your ORCID to the JOSS article, could you please provide it here asap?
