### 2022-09-29 pymedphys/pymedphys#1735 @pchlap
> if any of you would like to add your ORCID to the JOSS article, could you please provide it here asap?

0000-0002-6517-8745
