### 2022-09-29 pymedphys/pymedphys#1735 @nlambriICH
> if any of you would like to add your ORCID to the JOSS article, could you please provide it here asap?

0000-0001-8706-6480
