### 2022-09-29 1a6a675
Update lib/pymedphys/docs/contrib/info/release-guide.md
Co-authored-by: Simon Biggs <simon.biggs@radiotherapy.ai>
