### 2022-09-29 #4555 @osorensen
> > I'd note that a forthcoming 0.39.0 probably more closely matches what was reviewed, since virtually all of the changes in `main` made since the 0.38.0 release were made prior to commencement of the JOSS review, and we've made some small changes to address reviewer comments that will only be present in 0.39.0.
>
> @Matthew-Jennings, it makes the most sense that 0.39.0 is the reviewed version. No need to wait for a response here, move forward with 0.39.0 being the one, and have the changelog reflect that.

I agree. Just make sure that the version reported here corresponds to the Zenodo version.
