### 2022-09-29 pymedphys/pymedphys#1735 @SimonBiggs
Here's the Zenodo DOI:
https://doi.org/10.5281/zenodo.7126882

However, the paper in that release won't have these ORCIDs. If we want that, we'll need to release a v0.39.1 of PyMedPhys and have it make a new Zenodo release.
