### 2022-10-04 7b7f2c9
Update paper.md
Add @kingrpaul's ORCID
