### 2022-10-05 e0f0095
Add note and small edits to release guide.
