+++ requirements.txt
@@ -19,14 +19,14 @@ dbfread==2.0.7 ; python_version >= "3.8" and python_version < "3.11"
 decorator==5.1.1 ; python_version >= "3.8" and python_version < "3.11"
 dicompyler-core==0.5.5 ; python_version >= "3.8" and python_version < "3.11"
 entrypoints==0.4 ; python_version >= "3.8" and python_version < "3.11"
-fonttools==4.37.3 ; python_version >= "3.8" and python_version < "3.11"
+fonttools==4.37.4 ; python_version >= "3.8" and python_version < "3.11"
 fsspec==2022.8.2 ; python_version >= "3.8" and python_version < "3.11"
 gitdb==4.0.9 ; python_version >= "3.8" and python_version < "3.11"
 gitpython==3.1.27 ; python_version >= "3.8" and python_version < "3.11"
 greenlet==1.1.3 ; python_version >= "3.8" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "3.11"
 idna==3.4 ; python_version >= "3.8" and python_version < "3.11"
-imageio==2.22.0 ; python_version >= "3.8" and python_version < "3.11"
-importlib-metadata==4.12.0 ; python_version >= "3.8" and python_version < "3.11"
+imageio==2.22.1 ; python_version >= "3.8" and python_version < "3.11"
+importlib-metadata==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
 jaraco-classes==3.2.3 ; python_version >= "3.8" and python_version < "3.11"
 jeepney==0.8.0 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "linux"
 jinja2==3.1.2 ; python_version >= "3.8" and python_version < "3.11"
@@ -39,14 +39,14 @@ markupsafe==2.1.1 ; python_version >= "3.8" and python_version < "3.11"
 matplotlib==3.6.0 ; python_version >= "3.8" and python_version < "3.11"
 more-itertools==8.14.0 ; python_version >= "3.8" and python_version < "3.11"
 natsort==8.2.0 ; python_version >= "3.8" and python_version < "3.11"
-networkx==2.8.6 ; python_version >= "3.8" and python_version < "3.11"
+networkx==2.8.7 ; python_version >= "3.8" and python_version < "3.11"
 numpy==1.23.3 ; python_version >= "3.8" and python_version < "3.11"
 packaging==21.3 ; python_version >= "3.8" and python_version < "3.11"
 pandas==1.5.0 ; python_version >= "3.8" and python_version < "3.11"
 pillow==9.2.0 ; python_version >= "3.8" and python_version < "3.11"
 plotly==5.10.0 ; python_version >= "3.8" and python_version < "3.11"
 premailer==3.10.0 ; python_version >= "3.8" and python_version < "3.11"
-protobuf==3.20.2 ; python_version >= "3.8" and python_version < "3.11"
+protobuf==3.20.3 ; python_version >= "3.8" and python_version < "3.11"
 pyarrow==9.0.0 ; python_version >= "3.8" and python_version < "3.11"
 pycparser==2.21 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "linux"
 pydeck==0.8.0b3 ; python_version >= "3.8" and python_version < "3.11"
@@ -61,21 +61,21 @@ pyparsing==3.0.9 ; python_version >= "3.8" and python_version < "3.11"
[…]
