### 2022-10-05 1536413
add changelog records for 0.39.0 and 0.39.1
