### 2022-10-05 b61a3eb
fix release guide note
