### 2022-10-05 #4555 @Matthew-Jennings
Hi @osorensen,

We are just [completing a patch to include most of the authors ORCIDs in the paper](https://github.com/pymedphys/pymedphys/pull/1749). I expect that we'll have v0.39.1 archived on Zenodo archive by tomorrow.
