### 2022-10-06 d1fae92
switch back to using common package of pylibjpeg-libjpeg and pinned lower version to 1.3.2
