### 2022-10-06 132e549
bump to version 0.39.1 update CHANGELOG.md to notify of reference to pylibjpeg-libjpeg
