+++ requirements-all.txt
@@ -41,17 +41,17 @@ entrypoints==0.4 ; python_version >= "3.8" and python_version < "3.11"
 executing==1.1.0 ; python_version >= "3.8" and python_version < "3.11"
 fastjsonschema==2.16.2 ; python_version >= "3.8" and python_version < "3.11"
 filelock==3.8.0 ; python_version >= "3.8" and python_version < "3.11"
-fonttools==4.37.3 ; python_version >= "3.8" and python_version < "3.11"
+fonttools==4.37.4 ; python_version >= "3.8" and python_version < "3.11"
 fsspec==2022.8.2 ; python_version >= "3.8" and python_version < "3.11"
 gitdb==4.0.9 ; python_version >= "3.8" and python_version < "3.11"
 gitpython==3.1.27 ; python_version >= "3.8" and python_version < "3.11"
 greenlet==1.1.3 ; python_version >= "3.8" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "3.11"
 hypothesis==5.49.0 ; python_version >= "3.8" and python_version < "3.11"
-identify==2.5.5 ; python_version >= "3.8" and python_version < "3.11"
+identify==2.5.6 ; python_version >= "3.8" and python_version < "3.11"
 idna==3.4 ; python_version >= "3.8" and python_version < "3.11"
-imageio==2.22.0 ; python_version >= "3.8" and python_version < "3.11"
+imageio==2.22.1 ; python_version >= "3.8" and python_version < "3.11"
 imagesize==1.4.1 ; python_version >= "3.8" and python_version < "3.11"
-importlib-metadata==4.12.0 ; python_version >= "3.8" and python_version < "3.11"
+importlib-metadata==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
 importlib-resources==5.9.0 ; python_version >= "3.8" and python_version < "3.9"
 iniconfig==1.1.1 ; python_version >= "3.8" and python_version < "3.11"
 ipykernel==6.16.0 ; python_version >= "3.8" and python_version < "3.11"
@@ -89,7 +89,7 @@ mdit-py-plugins==0.2.8 ; python_version >= "3.8" and python_version < "3.11"
 mistune==0.8.4 ; python_version >= "3.8" and python_version < "3.11"
 more-itertools==8.14.0 ; python_version >= "3.8" and python_version < "3.11"
 mypy-extensions==0.4.3 ; python_version >= "3.8" and python_version < "3.11"
-mypy==0.981 ; python_version >= "3.8" and python_version < "3.11"
+mypy==0.982 ; python_version >= "3.8" and python_version < "3.11"
 myst-nb==0.13.2 ; python_version >= "3.8" and python_version < "3.11"
 myst-parser==0.15.2 ; python_version >= "3.8" and python_version < "3.11"
 natsort==8.2.0 ; python_version >= "3.8" and python_version < "3.11"
@@ -97,8 +97,8 @@ nbclient==0.5.13 ; python_version >= "3.8" and python_version < "3.11"
 nbconvert==6.5.4 ; python_version >= "3.8" and python_version < "3.11"
 nbdime==3.1.1 ; python_version >= "3.8" and python_version < "3.11"
 nbformat==5.6.1 ; python_version >= "3.8" and python_version < "3.11"
-nest-asyncio==1.5.5 ; python_version >= "3.8" and python_version < "3.11"
[…]
