### 2022-10-06 07afdc3
Make patient name anonymisation conformant
