+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pymedphys"
-version = "0.39.1"
+version = "0.40.0dev0"
 readme = "README.rst"
 description = "Medical Physics library"
 authors = [
