### 2022-10-06 fb3a7ff
0.39.1 release prep (#1750)
* switch back to using common package of pylibjpeg-libjpeg and pinned lower version to 1.3.2

* bump to version 0.39.1
update CHANGELOG.md to notify of reference to pylibjpeg-libjpeg
