+++ requirements-docs.txt
@@ -23,13 +23,13 @@ docutils==0.17.1 ; python_version >= "3.8" and python_version < "3.11"
 entrypoints==0.4 ; python_version >= "3.8" and python_version < "3.11"
 executing==1.1.0 ; python_version >= "3.8" and python_version < "3.11"
 fastjsonschema==2.16.2 ; python_version >= "3.8" and python_version < "3.11"
-fonttools==4.37.3 ; python_version >= "3.8" and python_version < "3.11"
+fonttools==4.37.4 ; python_version >= "3.8" and python_version < "3.11"
 gitdb==4.0.9 ; python_version >= "3.8" and python_version < "3.11"
 gitpython==3.1.27 ; python_version >= "3.8" and python_version < "3.11"
 greenlet==1.1.3 ; python_version >= "3.8" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "3.11"
 idna==3.4 ; python_version >= "3.8" and python_version < "3.11"
 imagesize==1.4.1 ; python_version >= "3.8" and python_version < "3.11"
-importlib-metadata==4.12.0 ; python_version >= "3.8" and python_version < "3.11"
+importlib-metadata==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
 importlib-resources==5.9.0 ; python_version >= "3.8" and python_version < "3.9"
 ipykernel==6.16.0 ; python_version >= "3.8" and python_version < "3.11"
 ipython-genutils==0.2.0 ; python_version >= "3.8" and python_version < "3.11"
@@ -63,8 +63,8 @@ nbclient==0.5.13 ; python_version >= "3.8" and python_version < "3.11"
 nbconvert==6.5.4 ; python_version >= "3.8" and python_version < "3.11"
 nbdime==3.1.1 ; python_version >= "3.8" and python_version < "3.11"
 nbformat==5.6.1 ; python_version >= "3.8" and python_version < "3.11"
-nest-asyncio==1.5.5 ; python_version >= "3.8" and python_version < "3.11"
-networkx==2.8.6 ; python_version >= "3.8" and python_version < "3.11"
+nest-asyncio==1.5.6 ; python_version >= "3.8" and python_version < "3.11"
+networkx==2.8.7 ; python_version >= "3.8" and python_version < "3.11"
 notebook==6.4.12 ; python_version >= "3.8" and python_version < "3.11"
 numpy==1.23.3 ; python_version >= "3.8" and python_version < "3.11"
 packaging==21.3 ; python_version >= "3.8" and python_version < "3.11"
@@ -88,7 +88,7 @@ pygments==2.13.0 ; python_version >= "3.8" and python_version < "3.11"
 pyparsing==3.0.9 ; python_version >= "3.8" and python_version < "3.11"
 pyrsistent==0.18.1 ; python_version >= "3.8" and python_version < "3.11"
 python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "3.11"
-pytz==2022.2.1 ; python_version >= "3.8" and python_version < "3.11"
+pytz==2022.4 ; python_version >= "3.8" and python_version < "3.11"
 pywin32==304 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "win32"
 pywinpty==2.0.8 ; python_version >= "3.8" and python_version < "3.11" and os_name == "nt"
 pyyaml==6.0 ; python_version >= "3.8" and python_version < "3.11"
[…]
