### 2022-10-06 434f021
update version to point to pre-release of next minor change reference from pymedphys specific libjpeg to common/pydicom pylibjpeg-libjpeg package, and specify minimum version of 1.3.2
