### 2022-10-06 #4555 @Matthew-Jennings
Hi @osorensen

Here is the Zenodo DOI for v0.39.1:
10.5281/zenodo.7151940

I think everything is in order. Thank you very much for your patience!
