+++ requirements-all.txt
@@ -131,9 +131,9 @@ pydata-sphinx-theme==0.8.1 ; python_version >= "3.8" and python_version < "3.11"
 pydeck==0.8.0b3 ; python_version >= "3.8" and python_version < "3.11"
 pydicom==2.3.0 ; python_version >= "3.8" and python_version < "3.11"
 pygments==2.13.0 ; python_version >= "3.8" and python_version < "3.11"
+pylibjpeg-libjpeg==1.3.2 ; python_version >= "3.8" and python_version < "3.11"
 pylinac==2.3.2 ; python_version >= "3.8" and python_version < "3.11"
 pylint==2.15.3 ; python_version >= "3.8" and python_version < "3.11"
-pymedphys-pylibjpeg-libjpeg==1.3.1 ; python_version >= "3.8" and python_version < "3.11"
 pympler==1.0.1 ; python_version >= "3.8" and python_version < "3.11"
 pymssql==2.2.5 ; python_version >= "3.8" and python_version < "3.11"
 pynetdicom==2.0.2 ; python_version >= "3.8" and python_version < "3.11"
