### 2022-10-06 d8f40ea
add logo to readme for Simon
