+++ README.rst
@@ -132,7 +132,7 @@ Maintainers
 .. _`Phillip Chlap`: https://github.com/pchlap
 
 
-|sjs| |rah|
+|rai| |sjs| |rah|
 
 Active contributors
 ****************************
@@ -175,6 +175,8 @@ Past contributors
 * `Paul King <https://github.com/kingrpaul>`_
 * `Jacob McAloney <https://github.com/JacobMcAloney>`_
 
+.. |rai| image:: https://github.com/pymedphys/pymedphys/raw/286deacdea2b3af9322796d413d0da6e1d8935a9/docs/logos/rai.png
+    :target: `Radiotherapy AI`_
 
 .. |rah| image:: https://github.com/pymedphys/pymedphys/raw/3f8d82fc3b53eb636a75336477734e39fa406110/docs/logos/gosa_200x200.png
     :target: `Royal Adelaide Hospital`_
