### 2022-10-06 09a8cc3
fix logos location within readme
