### 2022-10-06 #4555 @Matthew-Jennings
> Thanks @Matthew-Jennings. I the correct version now 0.39.2?

Yes, that is correct. Apologies for the rigmarole.
