+++ CHANGELOG.md
@@ -7,6 +7,10 @@ All notable changes to are documented here.
 This project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- No unreleased changes as of yet.
+
 ## [0.39.2]
 
 ### News around this release
@@ -1367,7 +1371,8 @@ pymedphys.zip_data_paths("mu-density-gui-e2e-data.zip", extract_directory=CWD)
 
 - Began keeping record of changes in `changelog.md`
 
-[unreleased]: https://github.com/pymedphys/pymedphys/compare/v0.39.1...main
+[unreleased]: https://github.com/pymedphys/pymedphys/compare/v0.39.2...main
+[0.39.2]: https://github.com/pymedphys/pymedphys/compare/v0.39.1...v0.39.2
 [0.39.1]: https://github.com/pymedphys/pymedphys/compare/v0.39.0...v0.39.1
 [0.39.0]: https://github.com/pymedphys/pymedphys/compare/v0.38.0...v0.39.0
 [0.38.0]: https://github.com/pymedphys/pymedphys/compare/v0.37.1...v0.38.0
