### 2022-10-06 #4555 @SimonBiggs
Out of interest, @osorensen, what hurdles remain before the final publishing?
