### 2022-10-06 #4555 @osorensen
> Out of interest, @osorensen, what hurdles remain before the final publishing?

The editor-in-chief will read through and make a final decision. Normally this takes less than a week.
