### 2022-10-06 #4555 @SimonBiggs
This is an amazing process you have set up. I'm very impressed :).

In the future we'd be keen to potentially submit follow up papers going with a focus on submodules within PyMedPhys as they expand and grow. I notice that right now the process is very one paper per one repo focused. How feasible would it be to in the future have other papers within the same repo?

Potentially might it be that we have the current paper.md file moved into a directory of papers, and the "paper being reviewed" could be the one at the top level of the repo while it is being reviewed? And then there comes a way of scoping the review to just one specific module that is being focused on for that paper...
