### 2022-10-06 #4555 @osorensen
Thanks, and yes, there are several cases where developers have submitted new papers when new functionality has been added to their software. One way to do this could be to open a new branch containing the paper, or as you suggest, having papers in a directory.
