### 2022-10-09 c348912
Add cities to affiliations Add "the" to RAH Expand USA
