### 2022-10-09 1a354f1
Add Wagga Wagga to CCP
