### 2022-10-09 cae8991
Apply SBs suggestions from code review
Co-authored-by: Simon Biggs <simon.biggs@radiotherapy.ai>
