### 2022-10-09 c556102
Add states to US affiliations
