+++ requirements.txt
@@ -22,8 +22,8 @@ entrypoints==0.4 ; python_version >= "3.8" and python_version < "3.11"
 fonttools==4.37.4 ; python_version >= "3.8" and python_version < "3.11"
 fsspec==2022.8.2 ; python_version >= "3.8" and python_version < "3.11"
 gitdb==4.0.9 ; python_version >= "3.8" and python_version < "3.11"
-gitpython==3.1.27 ; python_version >= "3.8" and python_version < "3.11"
-greenlet==1.1.3 ; python_version >= "3.8" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "3.11"
+gitpython==3.1.29 ; python_version >= "3.8" and python_version < "3.11"
+greenlet==1.1.3.post0 ; python_version >= "3.8" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "3.11"
 idna==3.4 ; python_version >= "3.8" and python_version < "3.11"
 imageio==2.22.1 ; python_version >= "3.8" and python_version < "3.11"
 importlib-metadata==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
@@ -36,7 +36,7 @@ keyring==23.9.3 ; python_version >= "3.8" and python_version < "3.11"
 kiwisolver==1.4.4 ; python_version >= "3.8" and python_version < "3.11"
 lxml==4.9.1 ; python_version >= "3.8" and python_version < "3.11"
 markupsafe==2.1.1 ; python_version >= "3.8" and python_version < "3.11"
-matplotlib==3.6.0 ; python_version >= "3.8" and python_version < "3.11"
+matplotlib==3.6.1 ; python_version >= "3.8" and python_version < "3.11"
 more-itertools==8.14.0 ; python_version >= "3.8" and python_version < "3.11"
 natsort==8.2.0 ; python_version >= "3.8" and python_version < "3.11"
 networkx==2.8.7 ; python_version >= "3.8" and python_version < "3.11"
@@ -71,7 +71,7 @@ requests==2.28.1 ; python_version >= "3.8" and python_version < "3.11"
 rich==12.6.0 ; python_version >= "3.8" and python_version < "3.11"
 scikit-image==0.19.3 ; python_version >= "3.8" and python_version < "3.11"
 scikit-learn==1.1.2 ; python_version >= "3.8" and python_version < "3.11"
-scipy==1.9.1 ; python_version >= "3.8" and python_version < "3.11"
+scipy==1.9.2 ; python_version >= "3.8" and python_version < "3.11"
 secretstorage==3.3.3 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "linux"
 semver==2.13.0 ; python_version >= "3.8" and python_version < "3.11"
 setuptools-scm==7.0.5 ; python_version >= "3.8" and python_version < "3.11"
@@ -92,7 +92,7 @@ tomlkit==0.11.5 ; python_version >= "3.8" and python_version < "3.11"
 toolz==0.12.0 ; python_version >= "3.8" and python_version < "3.11"
 tornado==6.2 ; python_version >= "3.8" and python_version < "3.11"
 tqdm==4.64.1 ; python_version >= "3.8" and python_version < "3.11"
-typing-extensions==4.3.0 ; python_version >= "3.8" and python_version < "3.11"
[…]
