+++ requirements-icom.txt
@@ -1,2 +1,2 @@
 numpy==1.23.3 ; python_version >= "3.8" and python_version < "3.11"
-typing-extensions==4.3.0 ; python_version >= "3.8" and python_version < "3.11"
+typing-extensions==4.4.0 ; python_version >= "3.8" and python_version < "3.11"
