+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "pymedphys"
-version = "0.40.0-dev1"
+version = "0.39.3"
 readme = "README.rst"
 description = "Medical Physics library"
 authors = [
