+++ CHANGELOG.md
@@ -11,14 +11,14 @@ This project adheres to
 
 - No unreleased changes as of yet.
 
-## [0.39.2]
+## [0.39.3]
 
 ### News around this release
 
 - This is the version of PyMedPhys reviewed for PyMedPhys' first publication
   within the [Journal of Open Source Software (JOSS)](https://joss.theoj.org/).
-  - Paper available for download at
-    [10.21105/joss.04555](https://doi.org/10.21105/joss.04555)
+
+## [0.39.2]
 
 ### Bug Fixes
 
@@ -1373,7 +1373,8 @@ pymedphys.zip_data_paths("mu-density-gui-e2e-data.zip", extract_directory=CWD)
 
 - Began keeping record of changes in `changelog.md`
 
-[unreleased]: https://github.com/pymedphys/pymedphys/compare/v0.39.2...main
+[unreleased]: https://github.com/pymedphys/pymedphys/compare/v0.39.3...main
+[0.39.2]: https://github.com/pymedphys/pymedphys/compare/v0.39.2...v0.39.3
 [0.39.2]: https://github.com/pymedphys/pymedphys/compare/v0.39.1...v0.39.2
 [0.39.1]: https://github.com/pymedphys/pymedphys/compare/v0.39.0...v0.39.1
 [0.39.0]: https://github.com/pymedphys/pymedphys/compare/v0.38.0...v0.39.0
