### 2022-10-15 14be5db
begin incorpoating econforge interpolator into gamma_shell
