### 2022-10-16 a272ed4
make interpolation optional. Lower bound = 2.1.0
