### 2022-10-16 212822c
safer, simpler CGrid (supports non-uniform axes)
