+++ requirements-all.txt
@@ -54,6 +54,7 @@ imagesize==1.4.1 ; python_version >= "3.8" and python_version < "3.11"
 importlib-metadata==5.0.0 ; python_version >= "3.8" and python_version < "3.11"
 importlib-resources==5.10.0 ; python_version >= "3.8" and python_version < "3.9"
 iniconfig==1.1.1 ; python_version >= "3.8" and python_version < "3.11"
+interpolation==2.2.4 ; python_version >= "3.8" and python_version < "3.11"
 ipykernel==6.16.0 ; python_version >= "3.8" and python_version < "3.11"
 ipython-genutils==0.2.0 ; python_version >= "3.8" and python_version < "3.11"
 ipython==8.5.0 ; python_version >= "3.8" and python_version < "3.11"
@@ -79,6 +80,7 @@ kiwisolver==1.4.4 ; python_version >= "3.8" and python_version < "3.11"
 latexcodec==2.0.1 ; python_version >= "3.8" and python_version < "3.11"
 lazy-object-proxy==1.7.1 ; python_version >= "3.8" and python_version < "3.11"
 linkify-it-py==1.0.3 ; python_version >= "3.8" and python_version < "3.11"
+llvmlite==0.39.1 ; python_version >= "3.8" and python_version < "3.11"
 lxml==4.9.1 ; python_version >= "3.8" and python_version < "3.11"
 markdown-it-py==1.1.0 ; python_version >= "3.8" and python_version < "3.11"
 markupsafe==2.1.1 ; python_version >= "3.8" and python_version < "3.11"
@@ -101,6 +103,7 @@ nest-asyncio==1.5.6 ; python_version >= "3.8" and python_version < "3.11"
 networkx==2.8.7 ; python_version >= "3.8" and python_version < "3.11"
 nodeenv==1.7.0 ; python_version >= "3.8" and python_version < "3.11"
 notebook==6.4.12 ; python_version >= "3.8" and python_version < "3.11"
+numba==0.56.3 ; python_version >= "3.8" and python_version < "3.11"
 numpy==1.23.3 ; python_version >= "3.8" and python_version < "3.11"
 packaging==21.3 ; python_version >= "3.8" and python_version < "3.11"
 pandas==1.5.0 ; python_version >= "3.8" and python_version < "3.11"
@@ -198,6 +201,7 @@ stevedore==4.0.0 ; python_version >= "3.8" and python_version < "3.11"
 streamlit-ace==0.1.1 ; python_version >= "3.8" and python_version < "3.11"
 streamlit==1.11.1 ; python_version >= "3.8" and python_version < "3.11"
 tabulate==0.9.0 ; python_version >= "3.8" and python_version < "3.11"
+tempita==0.5.2 ; python_version >= "3.8" and python_version < "3.11"
 tenacity==8.1.0 ; python_version >= "3.8" and python_version < "3.11"
 termcolor==2.0.1 ; python_version >= "3.8" and python_version < "3.11"
 terminado==0.16.0 ; python_version >= "3.8" and python_version < "3.11"
