+++ pyproject.toml
@@ -173,6 +173,7 @@ all = [
     "fsspec",
     "hypothesis",
     "imageio",
+    "interpolation",
     "ipython",
     "jupyter-book",
     "keyring",
@@ -271,6 +272,7 @@ user = [
     "dicompyler-core",
     "fsspec",
     "imageio",
+    "interpolation",
     "keyring",
     "matplotlib",
     "natsort",
