+++ lib/pymedphys/_imports/imports.py
@@ -9,6 +9,7 @@ import dateutil
 import dateutil.relativedelta
 import dbfread
 import interpolation
+import interpolation.splines
 import keyring
 import natsort
 import packaging
